+++ index.html
@@ -35,7 +35,6 @@
     <link rel="stylesheet" href="views/familyCalendar/calendar.css" />
     <link rel="stylesheet" href="views/familyPortraits/portraits.css" />
 
-
     <script src="views/fanChart/SVGfunctions.js"></script>
     <script src="views/fanChart/AhnenTafel.js"></script>
     <script src="views/fanChart/PeopleCollection.js"></script>
@@ -80,6 +79,11 @@
     <script src="views/cc7/js/date.format.js"></script>
     <script src="views/cc7/js/xlsx.full.min.js"></script>
     <script src="views/cc7/js/FileSaver.min.js"></script>
+    <script type="module" src="views/cc7/js/Settings.js"></script>
+    <script type="module" src="views/cc7/js/Utils.js"></script>
+    <script type="module" src="views/cc7/js/HierarchyView.js"></script>
+    <script type="module" src="views/cc7/js/LanceView.js"></script>
+    <script type="module" src="views/cc7/js/PeopleTable.js"></script>
     <script type="module" src="views/cc7/js/cc7.js"></script>
     <script type="module" src="views/cc7/cc7View.js"></script>
     <link rel="stylesheet" href="views/cc7/css/cc7.css" />
