+++ views/cc7/js/cc7.js
@@ -9,9 +9,11 @@
 import { theSourceRules } from "../../../lib/biocheck-api/src/SourceRules.js";
 import { BioCheckPerson } from "../../../lib/biocheck-api/src/BioCheckPerson.js";
 import { Biography } from "../../../lib/biocheck-api/src/Biography.js";
+import { PeopleTable } from "./PeopleTable.js";
+import { Settings } from "./Settings.js";
+import { Utils } from "./Utils.js";
 
 export class CC7 {
-    static APP_ID = "CC7";
     static LONG_LOAD_WARNING =
         "Loading 7 degrees may take a while, especially with Bio Check enabled (it can be 3 minutes or more) " +
         "so the default is set to 3. Feel free to change it. ";
@@ -225,614 +227,12 @@ export class CC7 {
 
     static GET_PEOPLE_LIMIT = 1000;
     static MAX_DEGREE = 7;
-    static HIDE = true;
-    static EXCEL = "xlsx";
-    static CSV = "csv";
 
-    static WANTED_NAME_PARTS = [
-        "Prefix",
-        "FirstName",
-        "MiddleNames",
-        "PreferredName",
-        "Nicknames",
-        "LastNameAtBirth",
-        "LastNameCurrent",
-        "Suffix",
-        "LastNameOther",
-    ];
-
-    // These images were obtained from either https://www.wikitree.com/wiki/Space:RTC_-_Resources
-    // or https://uxwing.com/. The latter states: "Exclusive collection of free icons download for
-    // commercial projects without attribution"
-    static dyIcons = [
-        {
-            value: "47px-RTC_-_Pictures.jpeg",
-            text: "IMG:views/cc7/images/47px-RTC_-_Pictures.jpeg",
-            width: 30,
-        },
-        {
-            value: "diedYoung.png",
-            text: "IMG:views/cc7/images/diedYoung.png",
-            width: 30,
-        },
-        {
-            value: "RTC_-_Pictures-6.png",
-            text: "IMG:views/cc7/images/RTC_-_Pictures-6.png",
-            width: 30,
-        },
-        { value: "br" },
-        {
-            value: "50px-Remember_the_Children-26.png",
-            text: "IMG:views/cc7/images/50px-Remember_the_Children-26.png",
-            width: 30,
-        },
-        {
-            value: "Remember_the_Children-21.png",
-            text: "IMG:views/cc7/images/Remember_the_Children-21.png",
-            width: 30,
-        },
-        {
-            value: "Remember_the_Children-27.png",
-            text: "IMG:views/cc7/images/Remember_the_Children-27.png",
-            width: 30,
-        },
-        { value: "br" },
-        {
-            value: "butterfly-icon.png",
-            text: "IMG:views/cc7/images/butterfly-icon.png",
-            width: 30,
-        },
-        {
-            value: "flower-plant-icon.png",
-            text: "IMG:views/cc7/images/flower-plant-icon.png",
-            width: 30,
-        },
-        {
-            value: "candle-light-icon.png",
-            text: "IMG:views/cc7/images/candle-light-icon.png",
-            width: 15,
-        },
-        { value: "br" },
-        {
-            value: "flower-rose-icon.png",
-            text: "IMG:views/cc7/images/flower-rose-icon.png",
-            width: 30,
-        },
-        {
-            value: "aids-ribbon-icon.png",
-            text: "IMG:views/cc7/images/aids-ribbon-icon.png",
-            width: 30,
-        },
-        {
-            value: "candle-light-color-icon.png",
-            text: "IMG:views/cc7/images/candle-light-color-icon.png",
-            width: 15,
-        },
-    ];
-
-    static optionsDef = {
-        viewClassName: "CC7View",
-        tabs: [
-            {
-                name: "icons",
-                label: "Died Young Icons",
-                hideSelect: true,
-                subsections: [{ name: "DiedYoungIcons", label: "Died young icons" }],
-                // comment: "",
-            },
-            {
-                name: "biocheck",
-                label: "Bio Check",
-                hideSelect: true,
-                subsections: [{ name: "BioCheckOptions", label: "Bio Check Options" }],
-                // comment: "",
-            },
-            {
-                name: "missingFamily",
-                label: "Missing Family",
-                hideSelect: true,
-                subsections: [{ name: "mfOptions", label: "Missing Family Options" }],
-                // comment: "",
-            },
-        ],
-        optionsGroups: [
-            {
-                tab: "icons",
-                subsection: "DiedYoungIcons",
-                category: "icons",
-                subcategory: "options",
-                options: [
-                    {
-                        optionName: "sect1",
-                        comment: "Icon to use for a child that died before age 5:",
-                        type: "br",
-                    },
-                    {
-                        optionName: "veryYoung",
-                        type: "radio",
-                        label: "",
-                        values: CC7.dyIcons,
-                        defaultValue: "47px-RTC_-_Pictures.jpeg",
-                    },
-                    {
-                        optionName: "sect2",
-                        comment: "Icon to use for a child that died before age 16:",
-                        type: "br",
-                    },
-                    {
-                        optionName: "young",
-                        type: "radio",
-                        label: "",
-                        values: CC7.dyIcons,
-                        defaultValue: "50px-Remember_the_Children-26.png",
-                    },
-                ],
-            },
-            {
-                tab: "biocheck",
-                subsection: "BioCheckOptions",
-                category: "biocheck",
-                subcategory: "options",
-                options: [
-                    {
-                        optionName: "bioComment",
-                        comment: "Enabling Bio Check only comes into effect at subsequent GET button clicks.",
-                        type: "br",
-                    },
-                    {
-                        optionName: "biocheckOn",
-                        type: "checkbox",
-                        label: "Enable Bio Check on all profiles",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "sect2",
-                        comment: "Don't highlight the following Bio Check Issues:",
-                        type: "br",
-                    },
-                    {
-                        optionName: "sectbr",
-                        label: "Section Messages",
-                        type: "br",
-                    },
-                    // --- Section Messages ---
-                    {
-                        optionName: "ackb4src",
-                        type: "checkbox",
-                        label: "Acknowledgements before Sources",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "acksubsec",
-                        type: "checkbox",
-                        label: "Acknowledgements not a section",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "adnotend",
-                        type: "checkbox",
-                        label: "Advance Directive not at end of profile",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "adsubsec",
-                        type: "checkbox",
-                        label: "Advance Directive not a section",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "bioempty",
-                        type: "checkbox",
-                        label: "Biography is empty",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "comnoend",
-                        type: "checkbox",
-                        label: "Comment with no ending",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "irefafter",
-                        type: "checkbox",
-                        label: "Inline <ref> tag after <references >",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "irefnoend",
-                        type: "checkbox",
-                        label: "Inline <ref> tag with no ending </ref> tag",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "linesbetween",
-                        type: "checkbox",
-                        label: "Lines between Sources and <references />",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "munsrc",
-                        type: "checkbox",
-                        label: "Marked unsourced",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "munsrcbut",
-                        type: "checkbox",
-                        label: "Marked unsourced but may have sources",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "maybeunsrc",
-                        type: "checkbox",
-                        label: "May be unsourced",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "misref",
-                        type: "checkbox",
-                        label: "Missing <references /> tag",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "misbio",
-                        type: "checkbox",
-                        label: "Missing Biography heading",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "missrc",
-                        type: "checkbox",
-                        label: "Missing Sources heading",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "multiref",
-                        type: "checkbox",
-                        label: "Multiple <references /> tag",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "multibio",
-                        type: "checkbox",
-                        label: "Multiple Biography headings",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "multisrc",
-                        type: "checkbox",
-                        label: "Multiple Sources headings",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "nodates",
-                        type: "checkbox",
-                        label: "Profile has no dates",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "badsrcs",
-                        type: "checkbox",
-                        label: "Sources not reliable/clearly identified",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "srcsubsec",
-                        type: "checkbox",
-                        label: "Sources not a section",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "spannoend",
-                        type: "checkbox",
-                        label: "Span with no end",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "stylebr",
-                        label: "Style Messages",
-                        type: "br",
-                    },
-                    // --- Style Messages ---
-                    {
-                        optionName: "adnonmem",
-                        type: "checkbox",
-                        label: "Advance Directive on non member",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "biohdb4",
-                        type: "checkbox",
-                        label: "Biography heading before ...",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "brnoend",
-                        type: "checkbox",
-                        label: "<BR without ending >",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "secempty",
-                        type: "checkbox",
-                        label: "Empty section",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "headb4bio",
-                        type: "checkbox",
-                        label: "Heading before Biography",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "horizb4bio",
-                        type: "checkbox",
-                        label: "Horizontal rule before Biography",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "bioemail",
-                        type: "checkbox",
-                        label: "Might contain email address",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "misrnb",
-                        type: "checkbox",
-                        label: "Missing Research Note box",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "navbb4",
-                        type: "checkbox",
-                        label: "Navigation Box before ...",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "navbnb4",
-                        type: "checkbox",
-                        label: "Navigation Box not before ...",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "nrhtml",
-                        type: "checkbox",
-                        label: "Not-recommended HTML tag(s)",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "pbb4",
-                        type: "checkbox",
-                        label: "Project Box before ...",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "pnb4bio",
-                        type: "checkbox",
-                        label: "Project ... not before Biography",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "rnbb4",
-                        type: "checkbox",
-                        label: "Research Note Box before ...",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "rnbnb4",
-                        type: "checkbox",
-                        label: "Research Note Box not before ...",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "rnbstatus",
-                        type: "checkbox",
-                        label: "Research Note Box status",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "stcknabio",
-                        type: "checkbox",
-                        label: "Sticker not after Biography heading",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "sumb4",
-                        type: "checkbox",
-                        label: "Summary Text before ...",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "ulb4bio",
-                        type: "checkbox",
-                        label: "Unexpected line before ...",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "ulmore",
-                        type: "checkbox",
-                        label: "Unexpected line ... more follow",
-                        defaultValue: 0,
-                    },
-                    {
-                        optionName: "wronglvl",
-                        type: "checkbox",
-                        label: "Wrong level heading",
-                        defaultValue: 0,
-                    },
-                ],
-            },
-            {
-                tab: "missingFamily",
-                subsection: "mfOptions",
-                category: "missingFamily",
-                subcategory: "options",
-                options: [
-                    {
-                        optionName: "mfComment",
-                        comment: "Include profiles with:",
-                        type: "br",
-                    },
-                    {
-                        optionName: "noParents",
-                        type: "checkbox",
-                        label: "No parents",
-                        defaultValue: 1,
-                    },
-                    {
-                        optionName: "oneParent",
-                        type: "checkbox",
-                        label: "One parent",
-                        defaultValue: 1,
-                    },
-                    {
-                        optionName: "noNoSpouses",
-                        type: "checkbox",
-                        label: 'The "No more spouses" box unchecked',
-                        defaultValue: 1,
-                    },
-                    {
-                        optionName: "noNoChildren",
-                        type: "checkbox",
-                        label: 'The "No more children" box unchecked',
-                        defaultValue: 1,
-                    },
-                    {
-                        optionName: "noChildren",
-                        type: "checkbox",
-                        label: 'No children and the "No more children" box unchecked',
-                        defaultValue: 1,
-                    },
-                ],
-            },
-        ],
-    };
-
-    // These message were obtained by grepping for 'Messages.push' in lib/biocheck-api/src/Biography.js
-    static reportMatches = [
-        // Section Messages
-        { opt: "ackb4src", re: /^Acknowledgements before/ },
-        { opt: "acksubsec", re: /^Acknowledgements subsection/ },
-        { opt: "adnotend", re: /^Advance Directive is not/ },
-        { opt: "adsubsec", re: /^Advance Directive subsection/ },
-        { opt: "bioempty", re: /^Biography is empty/ },
-        { opt: "comnoend", re: /^Comment with no ending/ },
-        { opt: "irefafter", re: /^Inline <ref> tag after/ },
-        { opt: "irefnoend", re: /^Inline <ref> tag with/ },
-        { opt: "linesbetween", re: /lines? between Sources/ },
-        { opt: "misref", re: /^Missing <references/ },
-        { opt: "misbio", re: /^Missing Biography/ },
-        { opt: "missrc", re: /^Missing Sources/ },
-        { opt: "multiref", re: /^Multiple <references/ },
-        { opt: "multibio", re: /^Multiple Biography/ },
-        { opt: "multisrc", re: /^Multiple Sources/ },
-        { opt: "munsrc", re: /^Profile is marked unsourced$/ },
-        { opt: "munsrcbut", re: /^Profile is marked unsourced but/ },
-        { opt: "maybeunsrc", re: /^Profile may be/ },
-        { opt: "nodates", re: /^Profile has no dates/ },
-        { opt: "badsrcs", re: /^Bio Check found sources/ },
-        { opt: "srcsubsec", re: /^Sources subsection/ },
-        { opt: "spannoend", re: /^Span with no end/ },
-        // Style Messages
-        { opt: "adnonmem", re: /^Advance Directive on a non/ },
-        { opt: "nrhtml", re: /^Biography contains HTML/ },
-        { opt: "brnoend", re: /^Biography has <BR/ },
-        { opt: "biohdb4", re: /^Biography heading before/ },
-        { opt: "bioemail", re: /^Biography may contain email/ },
-        { opt: "secempty", re: /^Empty Biography section/ },
-        { opt: "headb4bio", re: /^Heading or subheading/ },
-        { opt: "horizb4bio", re: /^Horizontal rule before/ },
-        { opt: "misrnb", re: /^Missing Research/ },
-        { opt: "navbb4", re: /^Navigation Box before/ },
-        { opt: "navbnb4", re: /^Navigation.+should be/ },
-        { opt: "pnb4bio", re: /^Project:.+should be/ },
-        { opt: "pbb4", re: /^Project Box before/ },
-        { opt: "rnbb4", re: /^Research Note Box before/ },
-        { opt: "rnbnb4", re: /^Research Note Box:.+should be/ },
-        { opt: "rnbstatus", re: /^Research Note Box:.+is.+status/ },
-        { opt: "stcknabio", re: /^Sticker.+should be after/ },
-        { opt: "sumb4", re: /^Summary Text before/ },
-        { opt: "ulmore", re: /^Unexpected line.+more/ },
-        { opt: "ulb4bio", re: /^Unexpected line before/ },
-        { opt: "wronglvl", re: /^Wrong level/ },
-    ];
-
-    static mustHighlight(bioReport) {
-        for (const [msg, subLines] of bioReport) {
-            let hadMatch = false;
-            for (const rm of CC7.reportMatches) {
-                if (CC7.currentSettings[`biocheck_options_${rm.opt}`] && msg.match(rm.re)) {
-                    hadMatch = true;
-                    break;
-                }
-            }
-            // if the current message did not match any of the flagged messages, it must be highligted
-            // so we don't even have to check the rest. If it did match, we still have to check if any
-            // of the other messages doesn't match
-            if (!hadMatch) return true;
-        }
-        return false;
-    }
-
-    static settingOptionsObj;
-    static currentSettings = {};
     static cancelLoadController;
-    static nextZLevel = 10000;
 
     constructor(selector, startId) {
         this.selector = selector;
-        const optionsJson = CC7.getCookie("w_diedYoung");
-        // console.log(`Retrieved options ${optionsJson}`);
-        if (optionsJson) {
-            // Restore the settings to what the user last saved
-            const opt = JSON.parse(optionsJson);
-            function optionWithDefault(theOption, theDefault) {
-                return typeof opt[theOption] == "undefined" ? theDefault : opt[theOption];
-            }
-            CC7.optionsDef.optionsGroups[0].options[1].defaultValue = optionWithDefault(
-                "icons_options_veryYoung",
-                "47px-RTC_-_Pictures.jpeg"
-            );
-            CC7.optionsDef.optionsGroups[0].options[3].defaultValue = optionWithDefault(
-                "icons_options_young",
-                "50px-Remember_the_Children-26.png"
-            );
-
-            CC7.optionsDef.optionsGroups[1].options[1].defaultValue = optionWithDefault(
-                "biocheck_options_biocheckOn",
-                0
-            );
-            CC7.optionsDef.optionsGroups[2].options[1].defaultValue = optionWithDefault(
-                "missingFamily_options_noParents",
-                1
-            );
-            CC7.optionsDef.optionsGroups[2].options[2].defaultValue = optionWithDefault(
-                "missingFamily_options_oneParent",
-                1
-            );
-            CC7.optionsDef.optionsGroups[2].options[3].defaultValue = optionWithDefault(
-                "missingFamily_options_noNoSpouses",
-                1
-            );
-            CC7.optionsDef.optionsGroups[2].options[4].defaultValue = optionWithDefault(
-                "missingFamily_options_noNoChildren",
-                1
-            );
-            CC7.optionsDef.optionsGroups[2].options[5].defaultValue = optionWithDefault(
-                "missingFamily_options_noChildren",
-                1
-            );
-            const msgOptions = CC7.optionsDef.optionsGroups[1].options;
-            for (const rm of CC7.reportMatches) {
-                for (const msgOpt of msgOptions) {
-                    if (msgOpt.optionName == rm.opt) {
-                        msgOpt.defaultValue = optionWithDefault(`biocheck_options_${rm.opt}`, 0);
-                        break;
-                    }
-                }
-            }
-        }
-        CC7.settingOptionsObj = new SettingsOptions.SettingsOptionsObject(CC7.optionsDef);
+        Settings.restoreSettings();
         $(selector).html(
             `<div id="cc7Container" class="cc7Table">
             <button
@@ -860,12 +260,12 @@ export class CC7 {
             <span id="settingsButton" title="Settings"><img src="./views/cc7/images/setting-icon.png" /></span>
             <span id="help" title="About this">?</span>
             </span>
-            ${CC7.settingOptionsObj.createdSettingsDIV}
+            ${Settings.getSettingsDiv()}
             <div id="explanation">${CC7.#helpText}</div>
             </div>`
         );
 
-        const cc7Degree = CC7.getCookie("w_cc7Degree");
+        const cc7Degree = Utils.getCookie("w_cc7Degree");
         if (cc7Degree && cc7Degree > 0 && cc7Degree <= CC7.MAX_DEGREE) {
             CC7.handleDegreeChange(cc7Degree);
         }
@@ -881,7 +281,7 @@ export class CC7 {
         $("#help")
             .off("click")
             .on("click", function () {
-                $("#explanation").css("z-index", `${CC7.nextZLevel++}`).slideToggle();
+                $("#explanation").css("z-index", `${Settings.getNextZLevel()}`).slideToggle();
             });
         $("#explanation")
             .off("dblclick")
@@ -907,14 +307,11 @@ export class CC7 {
                 CC7.toggleSettings();
             });
         $("#settingsDIV").draggable();
-
-        CC7.settingOptionsObj.buildPage();
-        CC7.settingOptionsObj.setActiveTab("icons");
-        CC7.currentSettings = CC7.settingOptionsObj.getDefaultOptions();
+        Settings.renderSettings();
         CC7.setInfoPanelMessage();
 
         $("#cancelLoad").off("click").on("click", CC7.cancelLoad);
-        $("#getDegreeButton").off("click").on("click", CC7.getDegreeAction);
+        $("#getDegreeButton").off("click").on("click", CC7.getOneDegreeOnly);
 
         $("#savePeople")
             .off("click")
@@ -936,9 +333,7 @@ export class CC7 {
         const loadWarning = CC7.firstTimeLoad ? CC7.LONG_LOAD_WARNING : "";
         wtViewRegistry.setInfoPanel(
             loadWarning +
-                `Bio Check is ${
-                    CC7.currentSettings["biocheck_options_biocheckOn"] ? "ENABLED" : "DISABLED"
-                } in settings.`
+                `Bio Check is ${Settings.current["biocheck_options_biocheckOn"] ? "ENABLED" : "DISABLED"} in settings.`
         );
         wtViewRegistry.showInfoPanel();
     }
@@ -954,7 +349,7 @@ export class CC7 {
     static toggleSettings() {
         const theDIV = document.getElementById("settingsDIV");
         if (theDIV.style.display == "none") {
-            theDIV.style.zIndex = `${CC7.nextZLevel++}`;
+            theDIV.style.zIndex = `${Settings.getNextZLevel()}`;
             theDIV.style.display = "block";
         } else {
             theDIV.style.display = "none";
@@ -962,11 +357,11 @@ export class CC7 {
     }
 
     static settingsChanged(e) {
-        // console.log("current settings:", CC7.currentSettings);
-        if (CC7.settingOptionsObj.hasSettingsChanged(CC7.currentSettings)) {
-            // console.log(`new settings: ${String(CC7.currentSettings)}`, CC7.currentSettings);
-            const veryYoungImg = CC7.imagePath(CC7.currentSettings["icons_options_veryYoung"]);
-            const youngImg = CC7.imagePath(CC7.currentSettings["icons_options_young"]);
+        // console.log("current settings:", Settings.current);
+        if (Settings.hasSettingsChanged()) {
+            // console.log(`new settings: ${String(Settings.current)}`, Settings.current);
+            const veryYoungImg = Utils.imagePath(Settings.current["icons_options_veryYoung"]);
+            const youngImg = Utils.imagePath(Settings.current["icons_options_young"]);
             $("img.diedVeryYoungImg").each(function () {
                 const it = $(this);
                 it.attr("src", veryYoungImg);
@@ -974,7 +369,7 @@ export class CC7 {
             $("img.diedYoungImg").each(function () {
                 $(this).attr("src", youngImg);
             });
-            if (!CC7.currentSettings["biocheck_options_biocheckOn"]) {
+            if (!Settings.current["biocheck_options_biocheckOn"]) {
                 $("td.bioIssue").each(function () {
                     $(this).off("click");
                     $(this).removeClass("bioIssue");
@@ -983,7 +378,7 @@ export class CC7 {
                 CC7.redoBioCheckFlags();
             }
             CC7.setInfoPanelMessage();
-            CC7.setCookie("w_diedYoung", JSON.stringify(CC7.currentSettings), { expires: 365 });
+            Utils.setCookie("w_diedYoung", JSON.stringify(Settings.current), { expires: 365 });
             if ($("#cc7Subset").val() == "missing-links") {
                 $("#cc7Subset").trigger("change");
             }
@@ -999,48 +394,11 @@ export class CC7 {
                 const person = window.people.get(+$row.attr("data-id"));
                 if (person) {
                     $td.removeClass("bioIssue bioIssue2").addClass(
-                        CC7.mustHighlight(person.bioCheckReport) ? "bioIssue" : "bioIssue2"
+                        Settings.mustHighlight(person.bioCheckReport) ? "bioIssue" : "bioIssue2"
                     );
                 }
             });
         });
-    }
-
-    static showMissingLinksButtons() {
-        $("#mlButtons").show();
-        if ($("#mlButtons").length == 0) {
-            const mlButtons = $(
-                "<div id=mlButtons>" +
-                    '<label><input id="mlNoParents" type="checkbox" class="mfCheckbox"> No parents</label>' +
-                    '<label><input id="mlOneParent" type="checkbox" class="mfCheckbox"> One parent</label>' +
-                    '<label><input id="mlNoNoSpouses" type="checkbox" class="mfCheckbox"> No "No more spouses"</label>' +
-                    '<label><input id="mlNoNoChildren" type="checkbox" class="mfCheckbox"> No "No more children"</label>' +
-                    '<label><input id="mlNoChildren" type="checkbox" class="mfCheckbox"> No children and no "No more children"</label>' +
-                    "</div>"
-            );
-            mlButtons.insertAfter($("#tableButtons"));
-            $(".mfCheckbox")
-                .off("change")
-                .on("change", function (e) {
-                    const id = $(this).attr("id");
-                    const optId = `#missingFamily_options_${id[2].toLowerCase() + id.substring(3)}`;
-                    $(optId).prop("checked", $(this).prop("checked"));
-                    $("#saveSettingsChanges").trigger("click");
-                });
-        }
-        CC7.setMissingLinkButtons();
-    }
-
-    static setMissingLinkButtons() {
-        $("#mlNoParents").prop("checked", CC7.currentSettings["missingFamily_options_noParents"]);
-        $("#mlOneParent").prop("checked", CC7.currentSettings["missingFamily_options_oneParent"]);
-        $("#mlNoNoSpouses").prop("checked", CC7.currentSettings["missingFamily_options_noNoSpouses"]);
-        $("#mlNoNoChildren").prop("checked", CC7.currentSettings["missingFamily_options_noNoChildren"]);
-        $("#mlNoChildren").prop("checked", CC7.currentSettings["missingFamily_options_noChildren"]);
-    }
-
-    static imagePath(fileName) {
-        return `./views/cc7/images/${fileName}`;
     }
 
     static handleDegreeChange(wantedDegree) {
@@ -1063,9 +421,9 @@ export class CC7 {
             select.value = newDegree;
         }
         // Set the cookie if required
-        theDegree = CC7.getCookie("w_cc7Degree");
+        theDegree = Utils.getCookie("w_cc7Degree");
         if (newDegree != theDegree) {
-            CC7.setCookie("w_cc7Degree", newDegree, { expires: 365 });
+            Utils.setCookie("w_cc7Degree", newDegree, { expires: 365 });
         }
     }
 
@@ -1086,2937 +444,13 @@ export class CC7 {
 
             // Close the popup with the highest z-index
             if (lastPopup) {
-                CC7.nextZLevel = highestZIndex;
+                Settings.setNextZLevel(highestZIndex);
                 lastPopup.slideUp();
             }
         }
     }
 
-    // Age functions
-    static getAge(birth, death) {
-        // must be date objects
-        let age = death.getFullYear() - birth.getFullYear();
-        let m = death.getMonth() - birth.getMonth();
-        if (m < 0 || (m === 0 && death.getDate() < birth.getDate())) {
-            age--;
-        }
-        return age;
-    }
-
-    static ageAtDeath(person, showStatus = true) {
-        // ages
-        let about = "";
-        let diedAged = "";
-        if (person?.BirthDate != undefined) {
-            if (
-                person["BirthDate"].length == 10 &&
-                person["BirthDate"] != "0000-00-00" &&
-                person["DeathDate"].length == 10 &&
-                person["DeathDate"] != "0000-00-00"
-            ) {
-                about = "";
-                const obDateBits = person["BirthDate"].split("-");
-                if (obDateBits[1] == "00") {
-                    obDateBits[1] = "06";
-                    obDateBits[2] = "15";
-                    about = "~";
-                } else if (obDateBits[2] == "00") {
-                    obDateBits[2] = "15";
-                    about = "~";
-                }
-                const odDateBits = person["DeathDate"].split("-");
-                if (odDateBits[1] == "00") {
-                    odDateBits[1] = "06";
-                    odDateBits[2] = "15";
-                    about = "~";
-                } else if (odDateBits[2] == "00") {
-                    odDateBits[2] = "15";
-                    about = "~";
-                }
-
-                diedAged = CC7.getAge(
-                    new Date(obDateBits[0], obDateBits[1], obDateBits[2]),
-                    new Date(odDateBits[0], odDateBits[1], odDateBits[2])
-                );
-            } else {
-                diedAged = "";
-            }
-        }
-        if (person?.DataStatus?.DeathDate) {
-            if (person.DataStatus.DeathDate == "after") {
-                about = ">";
-            }
-        }
-        if (diedAged == "" && diedAged != "0") {
-            return false;
-        } else if (showStatus == false) {
-            return diedAged;
-        } else {
-            return about + diedAged;
-        }
-    }
-    // End age functions
-
-    static showShakingTree(callback) {
-        if ($("#tree").length) {
-            $("#tree").slideDown("fast", "swing", callback);
-        } else {
-            const treeGIF = $("<img id='tree' src='./views/cc7/images/tree.gif'>");
-            treeGIF.appendTo("#cc7Container");
-            $("#tree").css({
-                "display": "block",
-                "margin": "auto",
-                "height": "50px",
-                "width": "50px",
-                "border-radius": "50%",
-                "border": "3px solid forestgreen",
-            });
-        }
-    }
-
-    static hideShakingTree() {
-        $("#tree").slideUp("fast");
-    }
-
-    static setOverflow(value) {
-        $("#view-container").css({
-            overflow: value,
-        });
-    }
-
-    static async addWideTableButton() {
-        $("#wideTableButton").show();
-
-        if ($("#wideTableButton").length == 0) {
-            const pTable = $(".peopleTable");
-            const wideTableButton = $(
-                "<div id='tableButtons'><button class='button small' id='wideTableButton'>Wide Table</button></div>"
-            );
-            wideTableButton.insertBefore(pTable);
-
-            $("#wideTableButton")
-                .off("click")
-                .on("click", function (e) {
-                    e.preventDefault();
-
-                    const dTable = $(".peopleTable").eq(0);
-                    if (CC7.getCookie("w_wideTable") == "1") {
-                        // Display a normal table
-                        CC7.setCookie("w_wideTable", 0, { expires: 365 });
-                        CC7.setOverflow("visible");
-                        dTable.removeClass("wide");
-                        $(this).text("Wide table");
-                        $("#peopleTable").attr("title", "");
-                        dTable.css({
-                            position: "relative",
-                            top: "0px",
-                            left: "0px",
-                        });
-                        dTable.draggable("disable");
-                    } else {
-                        // Display a wide table
-                        CC7.setCookie("w_wideTable", 1, { expires: 365 });
-                        CC7.setOverflow("auto");
-                        $(this).text("Normal Table");
-                        $("#peopleTable").attr("title", "Drag to scroll left or right");
-                        dTable.addClass("wide");
-                        dTable.find("th").each(function () {
-                            $(this).data("width", $(this).css("width"));
-                            $(this).css("width", "auto");
-                        });
-                        var isiPad = navigator.userAgent.match(/iPad/i) != null;
-                        if (isiPad) {
-                            if ($("#cc7Container").length) {
-                                dTable.draggable({
-                                    cursor: "grabbing",
-                                });
-                            }
-                        } else {
-                            if ($("#cc7Container").length) {
-                                dTable.draggable({
-                                    axis: "x",
-                                    cursor: "grabbing",
-                                });
-                            }
-                        }
-                        dTable.draggable("enable");
-                    }
-                });
-        }
-        if (CC7.getCookie("w_wideTable") == "1") {
-            CC7.setCookie("w_wideTable", 0, { expires: 365 });
-            $("#wideTableButton").trigger("click");
-        }
-    }
-
-    static htmlEntities(str) {
-        return String(str)
-            .replace(/&/g, "&amp;")
-            .replace(/</g, "&lt;")
-            .replace(/>/g, "&gt;")
-            .replace(/"/g, "&quot;")
-            .replace(/'/g, "&apos;");
-    }
-
-    static displayName(fPerson) {
-        let fName1 = "";
-        if (typeof fPerson["LongName"] != "undefined") {
-            if (fPerson["LongName"] != "") {
-                fName1 = fPerson["LongName"].replace(/\s\s/, " ");
-            }
-        }
-        let fName2 = "";
-        let fName4 = "";
-        if (typeof fPerson["MiddleName"] != "undefined") {
-            if (fPerson["MiddleName"] == "" && typeof fPerson["LongNamePrivate"] != "undefined") {
-                if (fPerson["LongNamePrivate"] != "") {
-                    fName2 = fPerson["LongNamePrivate"].replace(/\s\s/, " ");
-                }
-            }
-        } else {
-            if (typeof fPerson["LongNamePrivate"] != "undefined") {
-                if (fPerson["LongNamePrivate"] != "") {
-                    fName4 = fPerson["LongNamePrivate"].replace(/\s\s/, " ");
-                }
-            }
-        }
-
-        let fName3 = "";
-        const checks = [
-            "Prefix",
-            "FirstName",
-            "RealName",
-            "MiddleName",
-            "Nicknames",
-            "LastNameAtBirth",
-            "LastNameCurrent",
-            "Suffix",
-        ];
-        checks.forEach(function (dCheck) {
-            if (typeof fPerson["" + dCheck + ""] != "undefined") {
-                if (fPerson["" + dCheck + ""] != "" && fPerson["" + dCheck + ""] != null) {
-                    if (dCheck == "LastNameAtBirth") {
-                        if (fPerson["LastNameAtBirth"] != fPerson.LastNameCurrent) {
-                            fName3 += "(" + fPerson["LastNameAtBirth"] + ") ";
-                        }
-                    } else if (dCheck == "RealName") {
-                        if (typeof fPerson["FirstName"] != "undefined") {
-                        } else {
-                            fName3 += fPerson["RealName"] + " ";
-                        }
-                    } else if (dCheck == "Nicknames") {
-                        let nnamesSplit = fPerson.Nicknames.split(/,\s?/);
-                        let out = "";
-                        nnamesSplit.forEach(function (aNname, index) {
-                            nnamesSplit[index] = "“" + aNname + "”";
-                        });
-                        out += nnamesSplit.join(" ") + " ";
-                        fName3 += out;
-                    } else {
-                        fName3 += fPerson["" + dCheck + ""] + " ";
-                    }
-                }
-            }
-        });
-
-        const arr = [fName1, fName2, fName3, fName4];
-        var longest = arr.reduce(function (a, b) {
-            return a.length > b.length ? a : b;
-        });
-
-        const fName = longest;
-        let sName = fName;
-        if (fPerson["ShortName"]) {
-            sName = fPerson["ShortName"];
-        }
-
-        return [fName.trim(), sName.trim()];
-    }
-
-    static getOffset(el) {
-        const rect = el.getBoundingClientRect();
-        return {
-            left: rect.left + window.scrollX,
-            top: rect.top + window.scrollY,
-        };
-    }
-
-    static getTheYear(theDate, ev, person) {
-        if (!CC7.isOK(theDate)) {
-            if (ev == "Birth" || ev == "Death") {
-                theDate = person[ev + "DateDecade"];
-            }
-        }
-        let theDateM = theDate.match(/[0-9]{4}/);
-        if (CC7.isOK(theDateM)) {
-            return parseInt(theDateM[0]);
-        } else {
-            return false;
-        }
-    }
-
-    static dateToYMD(enteredDate) {
-        let enteredD;
-        if (enteredDate.match(/[0-9]{3,4}\-[0-9]{2}\-[0-9]{2}/)) {
-            enteredD = enteredDate;
-        } else {
-            let eDMonth = "00";
-            let eDYear = enteredDate.match(/[0-9]{3,4}/);
-            if (eDYear != null) {
-                eDYear = eDYear[0];
-            }
-            let eDDate = enteredDate.match(/\b[0-9]{1,2}\b/);
-            if (eDDate != null) {
-                eDDate = eDDate[0].padStart(2, "0");
-            }
-            if (eDDate == null) {
-                eDDate = "00";
-            }
-
-            if (enteredDate.match(/jan/i) != null) {
-                eDMonth = "01";
-            }
-            if (enteredDate.match(/feb/i) != null) {
-                eDMonth = "02";
-            }
-            if (enteredDate.match(/mar/i) != null) {
-                eDMonth = "03";
-            }
-            if (enteredDate.match(/apr/i) != null) {
-                eDMonth = "04";
-            }
-            if (enteredDate.match(/may/i) != null) {
-                eDMonth = "05";
-            }
-            if (enteredDate.match(/jun/i) != null) {
-                eDMonth = "06";
-            }
-            if (enteredDate.match(/jul/i) != null) {
-                eDMonth = "07";
-            }
-            if (enteredDate.match(/aug/i) != null) {
-                eDMonth = "08";
-            }
-            if (enteredDate.match(/sep/i) != null) {
-                eDMonth = "09";
-            }
-            if (enteredDate.match(/oct/i) != null) {
-                eDMonth = "10";
-            }
-            if (enteredDate.match(/nov/i) != null) {
-                eDMonth = "11";
-            }
-            if (enteredDate.match(/dec/i) != null) {
-                eDMonth = "12";
-            }
-            enteredD = eDYear + "-" + eDMonth + "-" + eDDate;
-        }
-        return enteredD;
-    }
-
-    static mapGender(gender, maleName, femaleName, neutralName) {
-        return gender == "Male" ? maleName : gender == "Female" ? femaleName : neutralName;
-    }
-
-    static capitalizeFirstLetter(string) {
-        return string.substring(0, 1).toUpperCase() + string.substring(1);
-    }
-
-    static #BMD_EVENTS = ["Birth", "Death", "Marriage"];
-
-    static getTimelineEvents(tPerson) {
-        const family = [tPerson].concat(tPerson.Parent, tPerson.Sibling, tPerson.Spouse, tPerson.Child);
-        const timeLineEvent = [];
-        const startDate = CC7.getTheYear(tPerson.BirthDate, "Birth", tPerson);
-
-        // Get all BMD events for each family member
-        family.forEach(function (aPerson) {
-            CC7.#BMD_EVENTS.forEach(function (ev) {
-                let evDate = "";
-                let evLocation;
-                if (aPerson[ev + "Date"]) {
-                    evDate = aPerson[ev + "Date"];
-                    evLocation = aPerson[ev + "Location"];
-                } else if (aPerson[ev + "DateDecade"]) {
-                    evDate = aPerson[ev + "DateDecade"];
-                    evLocation = aPerson[ev + "Location"];
-                }
-                if (ev == "Marriage") {
-                    const marriageData = tPerson.Marriage[aPerson.Id];
-                    if (marriageData && marriageData[ev + "Date"]) {
-                        evDate = marriageData[ev + "Date"];
-                        evLocation = marriageData[ev + "Location"];
-                    }
-                }
-                if (aPerson.Relation) {
-                    const theRelation = aPerson.Relation.replace(/s$/, "").replace(/ren$/, "");
-                    const gender = aPerson.Gender;
-                    if (theRelation == "Child") {
-                        aPerson.Relation = CC7.mapGender(gender, "son", "daughter", "child");
-                    } else if (theRelation == "Sibling") {
-                        aPerson.Relation = CC7.mapGender(gender, "brother", "sister", "sibling");
-                    } else if (theRelation == "Parent") {
-                        aPerson.Relation = CC7.mapGender(gender, "father", "mother", "parent");
-                    } else if (theRelation == "Spouse") {
-                        aPerson.Relation = CC7.mapGender(gender, "husband", "wife", "spouse");
-                    } else {
-                        aPerson.Relation = theRelation;
-                    }
-                }
-                if (evDate != "" && evDate != "0000" && CC7.isOK(evDate)) {
-                    let fName = aPerson.FirstName;
-                    if (!aPerson.FirstName) {
-                        fName = aPerson.RealName;
-                    }
-                    let bDate = aPerson.BirthDate;
-                    if (!aPerson.BirthDate) {
-                        bDate = aPerson.BirthDateDecade;
-                    }
-                    let mBio = aPerson.bio;
-                    if (!aPerson.bio) {
-                        mBio = "";
-                    }
-                    if (evLocation == undefined) {
-                        evLocation = "";
-                    }
-                    timeLineEvent.push({
-                        eventDate: evDate,
-                        location: evLocation,
-                        firstName: fName,
-                        LastNameAtBirth: aPerson.LastNameAtBirth,
-                        lastNameCurrent: aPerson.LastNameCurrent,
-                        birthDate: bDate,
-                        relation: aPerson.Relation,
-                        bio: mBio,
-                        evnt: ev,
-                        wtId: aPerson.Name,
-                    });
-                }
-            });
-            // Look for military events in bios
-            if (aPerson.bio) {
-                const tlTemplates = aPerson.bio.match(/\{\{[^]*?\}\}/gm);
-                if (tlTemplates != null) {
-                    const warTemplates = [
-                        "Creek War",
-                        "French and Indian War",
-                        "Iraq War",
-                        "Korean War",
-                        "Mexican-American War",
-                        "Spanish-American War",
-                        "The Great War",
-                        "US Civil War",
-                        "Vietnam War",
-                        "War in Afghanistan",
-                        "War of 1812",
-                        "World War II",
-                    ];
-                    tlTemplates.forEach(function (aTemp) {
-                        let evDate = "";
-                        let evLocation = "";
-                        let ev = "";
-                        let evDateStart = "";
-                        let evDateEnd = "";
-                        let evStart;
-                        let evEnd;
-                        aTemp = aTemp.replaceAll(/[{}]/g, "");
-                        const bits = aTemp.split("|");
-                        const templateTitle = bits[0].replaceAll(/\n/g, "").trim();
-                        bits.forEach(function (aBit) {
-                            const aBitBits = aBit.split("=");
-                            const aBitField = aBitBits[0].trim();
-                            if (aBitBits[1]) {
-                                const aBitFact = aBitBits[1].trim().replaceAll(/\n/g, "");
-                                if (warTemplates.includes(templateTitle) && CC7.isOK(aBitFact)) {
-                                    if (aBitField == "startdate") {
-                                        evDateStart = CC7.dateToYMD(aBitFact);
-                                        evStart = "joined " + templateTitle;
-                                    }
-                                    if (aBitField == "enddate") {
-                                        evDateEnd = CC7.dateToYMD(aBitFact);
-                                        evEnd = "left " + templateTitle;
-                                    }
-                                    if (aBitField == "enlisted") {
-                                        evDateStart = CC7.dateToYMD(aBitFact);
-                                        evStart = "enlisted for " + templateTitle.replace("american", "American");
-                                    }
-                                    if (aBitField == "discharged") {
-                                        evDateEnd = CC7.dateToYMD(aBitFact);
-                                        evEnd = "discharged from " + templateTitle.replace("american", "American");
-                                    }
-                                    if (aBitField == "branch") {
-                                        evLocation = aBitFact;
-                                    }
-                                }
-                            }
-                        });
-                        if (CC7.isOK(evDateStart)) {
-                            evDate = evDateStart;
-                            ev = evStart;
-                            timeLineEvent.push({
-                                eventDate: evDate,
-                                location: evLocation,
-                                firstName: aPerson.FirstName,
-                                LastNameAtBirth: aPerson.LastNameAtBirth,
-                                lastNameCurrent: aPerson.LastNameCurrent,
-                                birthDate: aPerson.BirthDate,
-                                relation: aPerson.Relation,
-                                bio: aPerson.bio,
-                                evnt: ev,
-                                wtId: aPerson.Name,
-                            });
-                        }
-                        if (CC7.isOK(evDateEnd)) {
-                            evDate = evDateEnd;
-                            ev = evEnd;
-                            timeLineEvent.push({
-                                eventDate: evDate,
-                                location: evLocation,
-                                firstName: aPerson.FirstName,
-                                LastNameAtBirth: aPerson.LastNameAtBirth,
-                                lastNameCurrent: aPerson.LastNameCurrent,
-                                birthDate: aPerson.BirthDate,
-                                relation: aPerson.Relation,
-                                bio: aPerson.bio,
-                                evnt: ev,
-                                wtId: aPerson.Name,
-                            });
-                        }
-                    });
-                }
-            }
-        });
-        return timeLineEvent;
-    }
-
-    static buildTimeline(tPerson, timelineEvents) {
-        const timelineTable = $(
-            `<div class='timeline' data-wtid='${tPerson.Name}'><w>↔</w><x>[ x ]</x><table class="timelineTable">` +
-                `<caption>Events in the life of ${tPerson.FirstName}'s family</caption>` +
-                "<thead><th class='tlDate'>Date</th><th class='tlBioAge'>Age</th>" +
-                "<th class='tlEventDescription'>Event</th><th class='tlEventLocation'>Location</th>" +
-                `</thead><tbody></tbody></table></div>`
-        );
-        let bpDead = false;
-        let bpDeadAge;
-        timelineEvents.forEach(function (aFact) {
-            // Add events to the table
-            const isEventForBioPerson = aFact.wtId == tPerson.Name;
-            const showDate = aFact.eventDate.replace("-00-00", "").replace("-00", "");
-            const tlDate = "<td class='tlDate'>" + showDate + "</td>";
-            let aboutAge = "";
-            let bpBdate = tPerson.BirthDate;
-            if (!tPerson.BirthDate) {
-                bpBdate = tPerson.BirthDateDecade.replace(/0s/, "5");
-            }
-            let hasBdate = true;
-            if (bpBdate == "0000-00-00") {
-                hasBdate = false;
-            }
-            const bpBD = CC7.getApproxDate(bpBdate);
-            const evDate = CC7.getApproxDate(aFact.eventDate);
-            const aPersonBD = CC7.getApproxDate(aFact.birthDate);
-            if (bpBD.Approx == true) {
-                aboutAge = "~";
-            }
-            if (evDate.Approx == true) {
-                aboutAge = "~";
-            }
-            const bpAgeAtEvent = CC7.getAge(new Date(bpBD.Date), new Date(evDate.Date));
-            let bpAge;
-            if (bpAgeAtEvent == 0) {
-                bpAge = "";
-            } else if (bpAgeAtEvent < 0) {
-                bpAge = `–${-bpAgeAtEvent}`;
-            } else {
-                bpAge = `${bpAgeAtEvent}`;
-            }
-            if (bpDead == true) {
-                const theDiff = parseInt(bpAgeAtEvent - bpDeadAge);
-                bpAge = "&#x1F397;+ " + theDiff;
-            }
-            let theBPAge;
-            if (aboutAge != "" && bpAge != "") {
-                theBPAge = "(" + bpAge + ")";
-            } else {
-                theBPAge = bpAge;
-            }
-            if (hasBdate == false) {
-                theBPAge = "";
-            }
-            const tlBioAge =
-                "<td class='tlBioAge'>" +
-                (aFact.evnt == "Death" && aFact.wtId == tPerson.Name ? "&#x1F397; " : "") +
-                theBPAge +
-                "</td>";
-            if (aFact.relation == undefined || isEventForBioPerson) {
-                aFact.relation = "";
-            }
-
-            let relation = aFact.relation.replace(/s$/, "");
-            const eventName = aFact.evnt.replaceAll(/Us\b/g, "US").replaceAll(/Ii\b/g, "II");
-
-            let fNames = aFact.firstName;
-            if (aFact.evnt == "marriage") {
-                fNames = tPerson.FirstName + " and " + aFact.firstName;
-                relation = "";
-            }
-            const tlFirstName = CC7.profileLink(aFact.wtId, fNames);
-            const tlEventLocation = "<td class='tlEventLocation'>" + aFact.location + "</td>";
-
-            if (aPersonBD.Approx == true) {
-                aboutAge = "~";
-            }
-            let aPersonAge = CC7.getAge(new Date(aPersonBD.Date), new Date(evDate.Date));
-            if (aPersonAge == 0 || aPersonBD.Date.match(/0000/) != null) {
-                aPersonAge = "";
-                aboutAge = "";
-            }
-            let theAge;
-            if (aboutAge != "" && aPersonAge != "") {
-                theAge = "(" + aPersonAge + ")";
-            } else {
-                theAge = aPersonAge;
-            }
-
-            let descr;
-            if (CC7.#BMD_EVENTS.includes(aFact.evnt)) {
-                descr =
-                    CC7.capitalizeFirstLetter(eventName) +
-                    " of " +
-                    (relation == "" ? relation : relation + ", ") +
-                    tlFirstName +
-                    (theAge == "" ? "" : ", " + theAge);
-            } else {
-                const who =
-                    relation == ""
-                        ? tlFirstName
-                        : CC7.capitalizeFirstLetter(relation) +
-                          " " +
-                          tlFirstName +
-                          (theAge == "" ? "" : ", " + theAge + ",");
-                descr = who + " " + eventName;
-            }
-
-            const tlEventDescription = "<td class='tlEventDescription'>" + descr + "</td>";
-
-            let classText = "";
-            if (isEventForBioPerson) {
-                classText += "BioPerson ";
-            }
-            classText += aFact.evnt + " ";
-            const tlTR = $(
-                "<tr class='" + classText + "'>" + tlDate + tlBioAge + tlEventDescription + tlEventLocation + "</tr>"
-            );
-            timelineTable.find("tbody").append(tlTR);
-            if (aFact.evnt == "Death" && aFact.wtId == tPerson.Name) {
-                bpDead = true;
-                bpDeadAge = bpAgeAtEvent;
-            }
-        });
-        return timelineTable;
-    }
-
-    static showTimeline(jqClicked) {
-        const theClickedRow = jqClicked.closest("tr");
-        const id = +theClickedRow.attr("data-id");
-        let tPerson = window.people.get(id);
-        const theClickedName = tPerson.Name;
-        const familyId = theClickedName.replace(" ", "_") + "_timeLine";
-        if ($(`#${familyId}`).length) {
-            $(`#${familyId}`).css("z-index", `${CC7.nextZLevel++}`).slideToggle();
-            return;
-        }
-
-        CC7.assignRelationshipsFor(tPerson);
-        const familyFacts = CC7.getTimelineEvents(tPerson);
-        // Sort the events
-        familyFacts.sort((a, b) => {
-            return a.eventDate.localeCompare(b.eventDate);
-        });
-        if (!tPerson.FirstName) {
-            tPerson.FirstName = tPerson.RealName;
-        }
-        // Make a table
-        const timelineTable = CC7.buildTimeline(tPerson, familyFacts, familyId);
-        timelineTable.attr("id", familyId);
-        CC7.showTable(jqClicked, timelineTable, 30, 30);
-    }
-
-    static showTable(jqClicked, theTable, lOffset, tOffset) {
-        // Attach the table to the container div
-        theTable.prependTo($("#cc7Container"));
-        theTable.draggable();
-        theTable.off("dblclick").on("dblclick", function () {
-            $(this).slideUp();
-        });
-
-        CC7.setOffset(jqClicked, theTable, lOffset, tOffset);
-        $(window).resize(function () {
-            if (theTable.length) {
-                CC7.setOffset(jqClicked, theTable, lOffset, tOffset);
-            }
-        });
-
-        theTable.css("z-index", `${CC7.nextZLevel++}`);
-        theTable.slideDown("slow");
-        theTable
-            .find("x")
-            .off("click")
-            .on("click", function () {
-                theTable.slideUp();
-            });
-        theTable
-            .find("w")
-            .off("click")
-            .on("click", function () {
-                theTable.toggleClass("wrap");
-            });
-    }
-
-    static setOffset(theClicked, elem, lOffset, tOffset) {
-        const theLeft = CC7.getOffset(theClicked[0]).left + lOffset;
-        elem.css({ top: CC7.getOffset(theClicked[0]).top + tOffset, left: theLeft });
-    }
-
-    static showBioCheckReport(jqClicked) {
-        const theClickedRow = jqClicked.closest("tr");
-        const id = +theClickedRow.attr("data-id");
-        let person = window.people.get(id);
-        if (typeof person.bioCheckReport == "undefined" || person.bioCheckReport.length == 0) {
-            return;
-        }
-        const theClickedName = person.Name;
-        const familyId = theClickedName.replace(" ", "_") + "_bioCheck";
-        if ($(`#${familyId}`).length) {
-            $(`#${familyId}`).css("z-index", `${CC7.nextZLevel++}`).slideToggle();
-            return;
-        }
-
-        const bioReportTable = CC7.getBioCheckReportTable(person);
-        bioReportTable.attr("id", familyId);
-        CC7.showTable(jqClicked, bioReportTable, 30, 30);
-    }
-
-    static getBioCheckReportTable(person) {
-        const issueWord = person.bioCheckReport.length == 1 ? "issue" : "issues";
-        const bioCheckTable = $(
-            `<div class='bioReport' data-wtid='${person.Name}'><w>↔</w><x>[ x ]</x><table class="bioReportTable">` +
-                `<caption>Bio Check found the following ${issueWord} with the biography of ${person.FirstName}</caption>` +
-                "<tbody><tr><td><ol></ol></td></tr></tbody></table></div>"
-        );
-
-        const ol = bioCheckTable.find("tbody ol");
-        for (const [msg, subLines] of person.bioCheckReport) {
-            let msgLI = $("<li></li>").text(msg);
-            if (subLines && subLines.length > 0) {
-                const subList = $("<ul></ul>");
-                for (const line of subLines) {
-                    subList.append($("<li></li>").text(line));
-                }
-                msgLI = msgLI.append(subList);
-            }
-            ol.append(msgLI);
-        }
-        return bioCheckTable;
-    }
-
-    static peopleToTable(kPeople) {
-        const personOfInterest = kPeople[0];
-        let disName = CC7.displayName(personOfInterest)[0];
-        if ($("#cc7Container").length) {
-            if (personOfInterest.MiddleName) {
-                disName = disName.replace(personOfInterest.MiddleName + " ", "");
-            }
-        }
-        const captionHTML = CC7.profileLink(CC7.htmlEntities(kPeople[0].Name), disName);
-        const kTable = $(
-            `<div class='familySheet'><w>↔</w><x>[ x ]</x><table><caption>${captionHTML}</caption>` +
-                "<thead><tr><th>Relation</th><th>Name</th><th>Birth Date</th><th>Birth Place</th><th>Death Date</th><th>Death Place</th></tr></thead>" +
-                "<tbody></tbody></table></div>"
-        );
-        kPeople.forEach(function (kPers) {
-            let rClass = "";
-            let isDecades = false;
-            kPers.RelationShow = kPers.Relation;
-            if (kPers.Relation == undefined || kPers.Active) {
-                kPers.Relation = "Sibling";
-                kPers.RelationShow = "";
-                rClass = "self";
-            }
-
-            let bDate;
-            if (kPers.BirthDate) {
-                bDate = kPers.BirthDate;
-            } else if (kPers.BirthDateDecade) {
-                bDate = kPers.BirthDateDecade.slice(0, -1) + "-00-00";
-                isDecades = true;
-            } else {
-                bDate = "0000-00-00";
-            }
-
-            let dDate;
-            if (kPers.DeathDate) {
-                dDate = kPers.DeathDate;
-            } else if (kPers.DeathDateDecade) {
-                if (kPers.DeathDateDecade == "unknown") {
-                    dDate = "0000-00-00";
-                } else {
-                    dDate = kPers.DeathDateDecade.slice(0, -1) + "-00-00";
-                }
-            } else {
-                dDate = "0000-00-00";
-            }
-
-            if (kPers.BirthLocation == null || kPers.BirthLocation == undefined) {
-                kPers.BirthLocation = "";
-            }
-
-            if (kPers.DeathLocation == null || kPers.DeathLocation == undefined) {
-                kPers.DeathLocation = "";
-            }
-
-            if (kPers.MiddleName == null) {
-                kPers.MiddleName = "";
-            }
-            let oName = CC7.displayName(kPers)[0];
-
-            if (kPers.Relation) {
-                kPers.Relation = kPers.Relation.replace(/s$/, "").replace(/ren$/, "");
-                if (rClass != "self") {
-                    kPers.RelationShow = kPers.Relation;
-                }
-            }
-            if (oName) {
-                let oBDate = CC7.ymdFix(bDate);
-                let oDDate = CC7.ymdFix(dDate);
-                if (isDecades == true) {
-                    oBDate = kPers.BirthDateDecade;
-                    if (oDDate != "") {
-                        oDDate = kPers.DeathDateDecade;
-                    }
-                }
-                const linkName = CC7.htmlEntities(kPers.Name);
-                const aLine = $(
-                    "<tr data-name='" +
-                        kPers.Name +
-                        "' data-birthdate='" +
-                        bDate.replaceAll(/\-/g, "") +
-                        "' data-relation='" +
-                        kPers.Relation +
-                        "' class='" +
-                        rClass +
-                        " " +
-                        kPers.Gender +
-                        "'><td>" +
-                        kPers.RelationShow +
-                        "</td><td>" +
-                        CC7.profileLink(linkName, oName) +
-                        "</td><td class='aDate'>" +
-                        oBDate +
-                        "</td><td>" +
-                        kPers.BirthLocation +
-                        "</td><td class='aDate'>" +
-                        oDDate +
-                        "</td><td>" +
-                        kPers.DeathLocation +
-                        "</td></tr>"
-                );
-
-                kTable.find("tbody").append(aLine);
-            }
-
-            if (kPers.Relation == "Spouse") {
-                let marriageDeets = "m.";
-                const marriageData = personOfInterest.Marriage[kPers.Id];
-                if (marriageData) {
-                    const dMdate = CC7.ymdFix(marriageData.MarriageDate);
-                    if (dMdate != "") {
-                        marriageDeets += " " + dMdate;
-                    }
-                    if (CC7.isOK(marriageData.MarriageLocation)) {
-                        marriageDeets += " " + marriageData.MarriageLocation;
-                    }
-                    if (marriageDeets != "m.") {
-                        const spouseLine = $(
-                            "<tr class='marriageRow " +
-                                kPers.Gender +
-                                "' data-spouse='" +
-                                kPers.Name +
-                                "'><td>&nbsp;</td><td colspan='3'>" +
-                                marriageDeets +
-                                "</td><td></td><td></td></tr>"
-                        );
-                        kTable.find("tbody").append(spouseLine);
-                    }
-                }
-            }
-        });
-        const rows = kTable.find("tbody tr");
-        rows.sort((a, b) => ($(b).data("birthdate") < $(a).data("birthdate") ? 1 : -1));
-        kTable.find("tbody").append(rows);
-
-        const familyOrder = ["Parent", "Sibling", "Spouse", "Child"];
-        familyOrder.forEach(function (relWord) {
-            kTable.find("tr[data-relation='" + relWord + "']").each(function () {
-                $(this).appendTo(kTable.find("tbody"));
-            });
-        });
-
-        kTable.find(".marriageRow").each(function () {
-            $(this).insertAfter(kTable.find("tr[data-name='" + $(this).data("spouse") + "']"));
-        });
-
-        return kTable;
-    }
-
-    static doFamilySheet(fPerson, theClicked) {
-        const theClickedName = fPerson.Name;
-        const familyId = theClickedName.replace(" ", "_") + "_family";
-        if ($(`#${familyId}`).length) {
-            $(`#${familyId}`).css("z-index", `${CC7.nextZLevel++}`).slideToggle();
-            return;
-        }
-
-        CC7.assignRelationshipsFor(fPerson);
-        const thisFamily = [fPerson].concat(fPerson.Parent, fPerson.Sibling, fPerson.Spouse, fPerson.Child);
-
-        const kkTable = CC7.peopleToTable(thisFamily);
-        kkTable.attr("id", familyId);
-        CC7.showTable(theClicked, kkTable, 30, 30);
-    }
-
-    static assignRelationshipsFor(person) {
-        person.Relation = undefined;
-        for (const rel of ["Parent", "Spouse", "Sibling", "Child"]) {
-            const relatives = person[rel];
-            if (relatives) {
-                for (const relative of relatives) {
-                    relative.Relation = rel;
-                }
-            }
-        }
-    }
-
-    static showFamilySheet(jqClicked) {
-        const theClickedRow = jqClicked.closest("tr");
-        const theClickedName = theClickedRow.attr("data-name");
-        const theClickedId = +theClickedRow.attr("data-id");
-
-        const aPeo = window.people.get(theClickedId);
-        if (aPeo?.Parent?.length > 0 || aPeo?.Child?.length > 0) {
-            CC7.doFamilySheet(aPeo, jqClicked);
-        } else {
-            console.log(`Calling getRelatives for ${theClickedName}`);
-            WikiTreeAPI.postToAPI({
-                appId: CC7.APP_ID,
-                action: "getRelatives",
-                getSpouses: "1",
-                getChildren: "1",
-                getParents: "1",
-                getSiblings: "1",
-                keys: theClickedName,
-            }).then((data) => {
-                // Construct this person so it conforms to the profiles retrieved using getPeople
-                const mPerson = CC7.convertToInternal(data[0].items[0].person);
-                window.people.set(+mPerson.Id, mPerson);
-                CC7.doFamilySheet(mPerson, jqClicked);
-            });
-        }
-    }
-
-    static convertToInternal(pData) {
-        const person = CC7.addFamilyToPerson(pData);
-        if (person.Parents) person.Parents = Object.keys(person.Parents);
-        if (person.Siblings) person.Siblings = Object.keys(person.Siblings);
-        if (person.Children) person.Children = Object.keys(person.Children);
-        person.Marriage = {};
-        if (person.Spouses) {
-            for (const sp of Object.values(person.Spouses)) {
-                person.Marriage[sp.Id] = {
-                    MarriageDate: sp.marriage_date,
-                    MarriageEndDate: sp.marriage_end_date,
-                    MarriageLocation: sp.marriage_location,
-                    DoNotDisplay: sp.do_not_display,
-                };
-            }
-        }
-        return person;
-    }
-
-    static addFamilyToPerson(person) {
-        person.Parent = CC7.getRels(person.Parents, "Parent");
-        person.Sibling = CC7.getRels(person.Siblings, "Sibling");
-        person.Spouse = CC7.getRels(person.Spouses, "Spouse");
-        person.Child = CC7.getRels(person.Children, "Child");
-        return person;
-    }
-
-    static getRels(rel, theRelation = false) {
-        const peeps = [];
-        if (typeof rel == undefined || rel == null) {
-            return peeps;
-        }
-        const pKeys = Object.keys(rel);
-        pKeys.forEach(function (pKey) {
-            const aPerson = rel[pKey];
-            if (theRelation != false) {
-                aPerson.Relation = theRelation;
-            }
-            peeps.push(aPerson);
-        });
-
-        return peeps;
-    }
-
-    static getDateFormat(fbds) {
-        let fullDateFormat = "j M Y";
-
-        let dateFormat;
-        const savedDateFormat = CC7.getCookie("w_dateFormat");
-        if (savedDateFormat) {
-            dateFormat = savedDateFormat;
-        } else {
-            dateFormat = 0;
-            fullDateFormat = "M j, Y";
-        }
-        if (dateFormat == 1) {
-            fullDateFormat = "j M Y";
-        }
-        if (dateFormat == 2) {
-            fullDateFormat = "F j, Y";
-        } else if (dateFormat == 3) {
-            fullDateFormat = "j F Y";
-        }
-
-        let fbd;
-        if (fbds[1] != "00" && fbds[2] != "00" && fbds[0] != "00") {
-            // month is zero-indexed(!)
-            const fbdsDate = new Date(fbds[0], parseInt(fbds[1]) - 1, fbds[2]);
-            fbd = fbdsDate.format("j M Y");
-            if (dateFormat > 0) {
-                fbd = fbdsDate.format(fullDateFormat);
-            }
-        } else if (fbds[1] != "00" && fbds[2] == "00" && fbds[0] != "00") {
-            // month is zero-indexed(!)
-            const fbdsDate = new Date(fbds[0], parseInt(fbds[1]) - 1, 1);
-            fbd = fbdsDate.format("M Y");
-            if (dateFormat > 1) {
-                fbd = fbdsDate.format("F Y");
-            }
-        } else if (fbds[1] != "00" && fbds[2] == "00") {
-            // month is zero-indexed(!)
-            const fbdsDate = new Date(fbds[0], parseInt(fbds[1]) - 1, 1);
-            fbd = fbdsDate.format("M Y");
-            if (dateFormat > 1) {
-                fbd = fbdsDate.format("F Y");
-            }
-        } else {
-            // month is zero-indexed(!)
-            const fbdsDate = new Date(fbds[0], 0, 1);
-            fbd = fbdsDate.format("Y");
-        }
-        return fbd;
-    }
-
-    static bdDatesStatus(person) {
-        let statusChoice = "symbols";
-        let abbr = false;
-        if ($("input[name='statusChoice'][value='abbreviations']").prop("checked") == true) {
-            statusChoice = "abbreviations";
-            abbr = true;
-        }
-
-        var bdStatus = "";
-        var ddStatus = "";
-        if (typeof person["DataStatus"] != "undefined") {
-            if (person["BirthDate"] != "0000-00-00") {
-                if (person["DataStatus"]["BirthDate"] != "") {
-                    if (person["DataStatus"]["BirthDate"] == "guess") {
-                        bdStatus = "~";
-                        if (abbr) {
-                            bdStatus = "abt. ";
-                        }
-                    } else if (person["DataStatus"]["BirthDate"] == "before") {
-                        bdStatus = "<";
-                        if (abbr) {
-                            bdStatus = "bef. ";
-                        }
-                    } else if (person["DataStatus"]["BirthDate"] == "after") {
-                        bdStatus = ">";
-                        if (abbr) {
-                            bdStatus = "aft. ";
-                        }
-                    }
-                }
-            }
-        }
-        if (typeof person["DataStatus"] != "undefined") {
-            if (person["DeathDate"] != "0000-00-00") {
-                if (person["DataStatus"]["DeathDate"] != "") {
-                    if (person["DataStatus"]["DeathDate"] == "guess") {
-                        ddStatus = "~";
-                        if (abbr) {
-                            ddStatus = "abt. ";
-                        }
-                    } else if (person["DataStatus"]["DeathDate"] == "before") {
-                        ddStatus = "<";
-                        if (abbr) {
-                            ddStatus = "bef. ";
-                        }
-                    } else if (person["DataStatus"]["DeathDate"] == "after") {
-                        ddStatus = ">";
-                        if (abbr) {
-                            ddStatus = "aft. ";
-                        }
-                    }
-                }
-            }
-        }
-        return [bdStatus, ddStatus];
-    }
-
-    static ymdFix(date) {
-        let outDate;
-        if (date == undefined || date == "") {
-            outDate = "";
-        } else {
-            const dateBits1 = date.split(" ");
-            if (dateBits1[2]) {
-                const sMonths = ["Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"];
-                const dMonth = date.match(/[A-z]+/i);
-                let dMonthNum;
-                if (dMonth != null) {
-                    sMonths.forEach(function (aSM, i) {
-                        if (
-                            dMonth[0].toLowerCase() == aSM.toLowerCase() ||
-                            dMonth[0].toLowerCase() == aSM + ".".toLowerCase()
-                        ) {
-                            dMonthNum = (i + 1).toString().padStart(2, "0");
-                        }
-                    });
-                }
-                const dDate = date.match(/\b[0-9]{1,2}\b/);
-                const dDateNum = dDate[0];
-                const dYear = date.match(/\b[0-9]{4}\b/);
-                const dYearNum = dYear[0];
-                outDate = dYearNum + "-" + dMonthNum + "-" + dDateNum;
-            } else {
-                const dateBits = date.split("-");
-                outDate = date;
-                if (dateBits[1] == "00" && dateBits[2] == "00") {
-                    if (dateBits[0] == "0000") {
-                        outDate = "";
-                    } else {
-                        outDate = dateBits[0];
-                    }
-                }
-            }
-        }
-        return outDate;
-    }
-
-    static isOK(thing) {
-        const excludeValues = [
-            "",
-            null,
-            "null",
-            "0000-00-00",
-            "unknown",
-            "Unknown",
-            "undefined",
-            undefined,
-            "0000",
-            "0",
-            0,
-            -1,
-            "-1",
-        ];
-        if (!excludeValues.includes(thing)) {
-            if (CC7.isNumeric(thing)) {
-                return true;
-            } else {
-                if (jQuery.type(thing) === "string") {
-                    const nanMatch = thing.match(/NaN/);
-                    if (nanMatch == null) {
-                        return true;
-                    } else {
-                        return false;
-                    }
-                } else {
-                    return false;
-                }
-            }
-        } else {
-            return false;
-        }
-    }
-
-    static isNumeric(n) {
-        return !isNaN(parseFloat(n)) && isFinite(n);
-    }
-
-    static location2ways(locationText) {
-        const alSplit = locationText.split(",");
-        const alSplit2 = alSplit.map((string) => string.trim());
-        const s2b = alSplit2.join(", ");
-        const b2s = alSplit2.reverse().join(", ");
-        return [s2b, b2s];
-    }
-
-    static secondarySort3(aList, dataThing1, dataThing2, isText = 0, reverse = 0) {
-        let lastOne = "Me";
-        let tempArr = [lastOne];
-
-        const rows = aList.find("li");
-        rows.each(function (index) {
-            if ($(this).data(dataThing1) == lastOne) {
-                tempArr.push($(this));
-            } else {
-                tempArr.sort(function (a, b) {
-                    if (isText == 1) {
-                        if (reverse == 0) {
-                            return $(a).data(dataThing2).localeCompare($(b).data(dataThing2));
-                        } else {
-                            return $(b).data(dataThing2).localeCompare($(a).data(dataThing2));
-                        }
-                    } else {
-                        if (reverse == 0) {
-                            return $(a).data(dataThing2) - $(b).data(dataThing2);
-                        } else {
-                            return $(b).data(dataThing2) - $(a).data(dataThing2);
-                        }
-                    }
-                });
-
-                tempArr.forEach(function (item) {
-                    if (lastOne != "Me") {
-                        item.insertBefore(rows.eq(index));
-                    }
-                });
-                tempArr = [$(this)];
-            }
-            lastOne = $(this).data(dataThing1);
-        });
-    }
-
-    static fillLocations(rows, order) {
-        rows.each(function () {
-            $(this)
-                .find("td.birthlocation")
-                .text($(this).attr("data-birthlocation" + order));
-            $(this)
-                .find("td.deathlocation")
-                .text($(this).attr("data-deathlocation" + order));
-        });
-        return rows;
-    }
-
-    static sortByThis(el) {
-        const aTable = $("#peopleTable");
-        el.off("click").on("click", function () {
-            let sorter = el.attr("id");
-            let rows = aTable.find("tbody tr");
-            if (sorter == "birthlocation" || sorter == "deathlocation") {
-                if (sorter == "birthlocation") {
-                    if (el.attr("data-order") == "s2b") {
-                        sorter = "birthlocation-reversed";
-                        el.attr("data-order", "b2s");
-                        rows = CC7.fillLocations(rows, "-reversed");
-                    } else {
-                        el.attr("data-order", "s2b");
-                        rows = CC7.fillLocations(rows, "");
-                    }
-                } else if (sorter == "deathlocation") {
-                    if (el.attr("data-order") == "s2b") {
-                        sorter = "deathlocation-reversed";
-                        el.attr("data-order", "b2s");
-                        rows = CC7.fillLocations(rows, "-reversed");
-                    } else {
-                        el.attr("data-order", "s2b");
-                        rows = CC7.fillLocations(rows, "");
-                    }
-                }
-                rows.sort(function (a, b) {
-                    if ($(b).data(sorter) == "") {
-                        return true;
-                    }
-                    return $(a).data(sorter).localeCompare($(b).data(sorter));
-                });
-            } else if (isNaN(rows.data(sorter))) {
-                if (el.attr("data-order") == "asc") {
-                    rows.sort(function (a, b) {
-                        if ($(a).data(sorter) == "") {
-                            return true;
-                        }
-                        return $(b).data(sorter).toString().localeCompare($(a).data(sorter));
-                    });
-                    el.attr("data-order", "desc");
-                } else {
-                    rows.sort(function (a, b) {
-                        if ($(b).data(sorter) == "") {
-                            return true;
-                        }
-                        return $(a).data(sorter).toString().localeCompare($(b).data(sorter));
-                    });
-                    el.attr("data-order", "asc");
-                }
-            } else {
-                if (el.attr("data-order") == "asc") {
-                    rows.sort((a, b) => ($(b).data(sorter) > $(a).data(sorter) ? 1 : -1));
-                    el.attr("data-order", "desc");
-                } else {
-                    rows.sort((a, b) => ($(a).data(sorter) > $(b).data(sorter) ? 1 : -1));
-                    el.attr("data-order", "asc");
-                }
-            }
-            aTable.find("tbody").append(rows);
-            rows.each(function () {
-                const toBottom = ["", "00000000"];
-                if (toBottom.includes(el.data(sorter))) {
-                    aTable.find("tbody").append(el);
-                }
-            });
-            aTable.find("tr.main").prependTo(aTable.find("tbody"));
-        });
-    }
-
-    static profileLink(wtRef, text) {
-        return wtRef.startsWith("Private")
-            ? text
-            : `<a target='_blank' href='https://www.wikitree.com/wiki/${wtRef}'>${text}</a>`;
-    }
-
-    static changesLink(wtRef, text) {
-        return wtRef.startsWith("Private")
-            ? text
-            : `<a target='_blank' href='https://www.wikitree.com/index.php?title=Special:NetworkFeed&who=${wtRef}'>${text}</a>`;
-    }
-
-    static ONE_DEGREE = true;
-    static tableCaption(id, theDegree, isOneDegree) {
-        let caption = "";
-        if (isOneDegree) {
-            caption = `Degree ${theDegree} connected people for ${id}`;
-        } else {
-            const person = window.people.get(id);
-            caption = person ? `CC${theDegree} for ${new PersonName(person).withParts(CC7.WANTED_NAME_PARTS)}` : "";
-        }
-        return caption;
-    }
-
-    // From https://github.com/wikitree/wikitree-api/blob/main/getProfile.md :
-    // Privacy_IsPrivate            True if Privacy = 20
-    // Privacy_IsPublic             True if Privacy = 50
-    // Privacy_IsOpen               True if Privacy = 60
-    // Privacy_IsAtLeastPublic      True if Privacy >= 50
-    // Privacy_IsSemiPrivate        True if Privacy = 30-40
-    // Privacy_IsSemiPrivateBio     True if Privacy = 30
-    static PRIVACY_LEVELS = new Map([
-        [60, { title: "Privacy: Open", img: "./views/cc7/images/privacy_open.png" }],
-        [50, { title: "Public", img: "./views/cc7/images/privacy_public.png" }],
-        [40, { title: "Private with Public Bio and Tree", img: "./views/cc7/images/privacy_public-tree.png" }],
-        [35, { title: "Private with Public Tree", img: "./views/cc7/images/privacy_privacy35.png" }],
-        [30, { title: "Public Bio", img: "./views/cc7/images/privacy_public-bio.png" }],
-        [20, { title: "Private", img: "./views/cc7/images/privacy_private.png" }],
-        [10, { title: "Unlisted", img: "./views/cc7/images/privacy_unlisted.png" }],
-    ]);
-
-    static async addPeopleTable(caption) {
-        $("#savePeople").show();
-        const sortTitle = "title='Click to sort'";
-        const aCaption = `<caption>${caption}</caption>`;
-        const degreeTH = `<th id='degree' ${sortTitle}>°</th>`;
-        const createdTH = `<th id='created' ${sortTitle} data-order='asc'>Created</th>`;
-        const touchedTH = `<th id='touched' ${sortTitle} data-order='asc'>Modified</th>`;
-        const parentsNum = "<th id='parent' title='Parents. Click to sort.' data-order='desc'>Par.</th>";
-        const siblingsNum = "<th id='sibling' title='Siblings. Click to sort.' data-order='desc'>Sib.</th>";
-        const spousesNum = "<th id='spouse' title='Spouses. Click to sort.' data-order='desc'>Sp.</th>";
-        const childrenNum = "<th id='child' title='Children. Click to sort.' data-order='desc'>Ch.</th>";
-        const ageAtDeathCol = "<th id='age-at-death' title='Age at Death. Click to sort.'  data-order='desc'>Age</th>";
-        const bioCheck = CC7.currentSettings["biocheck_options_biocheckOn"];
-
-        const aTable = $(
-            "<table id='peopleTable' class='peopleTable'>" +
-                aCaption +
-                `<thead><tr><th title='Privacy${bioCheck ? "/BioCheck" : ""}'>P${
-                    bioCheck ? "/B" : ""
-                }</th><th></th><th></th>` +
-                degreeTH +
-                parentsNum +
-                siblingsNum +
-                spousesNum +
-                childrenNum +
-                `<th data-order='' id='firstname' ${sortTitle}>Given Name(s)</th>` +
-                `<th data-order='' id='lnab' ${sortTitle}>Last Name at Birth</th>` +
-                `<th data-order='' id='lnc' ${sortTitle}>Current Last Name</th>` +
-                `<th data-order='' id='birthdate' ${sortTitle}>Birth Date</th>` +
-                `<th data-order='' id='birthlocation' ${sortTitle}>Birth Place</th>` +
-                `<th data-order='' id='deathdate' ${sortTitle}>Death Date</th>` +
-                `<th data-order='' id='deathlocation' ${sortTitle}>Death Place</th>` +
-                ageAtDeathCol +
-                `<th data-order='' id='manager' ${sortTitle}>Manager</th>` +
-                createdTH +
-                touchedTH +
-                "</tr></thead><tbody></tbody></table>"
-        );
-
-        if ($(".peopleTable").length) {
-            $(".peopleTable").eq(0).replaceWith(aTable);
-        } else {
-            aTable.appendTo($("#cc7Container"));
-        }
-
-        function sortIdsByDegreeAndBirthDate(keys) {
-            const collator = new Intl.Collator();
-            return [...keys]
-                .map((k) => {
-                    const mPerson = window.people.get(k);
-                    if (typeof mPerson.fixedBirthDate === "undefined") {
-                        CC7.fixBirthDate(mPerson);
-                    }
-                    const degree = +mPerson.Meta.Degrees;
-                    return [+k, degree < 0 ? 100 : degree, mPerson.fixedBirthDate];
-                })
-                .sort((a, b) => {
-                    if (a[1] == b[1]) {
-                        return collator.compare(a[2], b[2]);
-                    } else {
-                        return a[1] - b[1];
-                    }
-                });
-        }
-
-        const subset = $("#cc7Subset").val();
-        for (let [id, , birthDate] of sortIdsByDegreeAndBirthDate(window.people.keys())) {
-            const mPerson = window.people.get(id);
-            if (mPerson.Hide) continue;
-            switch (subset) {
-                case "above":
-                    if (!mPerson.isAbove) continue; // hide
-                    break;
-
-                case "below":
-                    if (mPerson.isAbove) continue; // hide
-                    break;
-
-                case "ancestors":
-                    if (mPerson.isAncestor) break; // show
-                    continue;
-
-                case "descendants":
-                    if (typeof mPerson.isAncestor != "undefined" && !mPerson.isAncestor) break; // show
-                    continue;
-
-                case "missing-links":
-                    if (isMissingFamily(mPerson)) break; // show
-                    continue; // else hide
-
-                default:
-                    break;
-            }
-            function isMissingFamily(person) {
-                if (person.LastNameAtBirth == "Private") return false;
-                let val = false;
-                if (CC7.currentSettings["missingFamily_options_noNoChildren"]) {
-                    // no more children flag is not set
-                    val = person.NoChildren != 1;
-                }
-                if (!val && CC7.currentSettings["missingFamily_options_noNoSpouses"]) {
-                    // no more spouses flag is not set
-                    val = person.DataStatus.Spouse != "blank";
-                }
-                if (!val && CC7.currentSettings["missingFamily_options_noParents"]) {
-                    // no father or mother
-                    val = !person.Father && !person.Mother;
-                }
-                if (!val && CC7.currentSettings["missingFamily_options_noChildren"]) {
-                    // no more children flag is not set and they don't have any children
-                    val = person.NoChildren != 1 && (!person.Child || person.Child.length == 0);
-                }
-                if (!val && CC7.currentSettings["missingFamily_options_oneParent"]) {
-                    // at least one parent missing
-                    val = (person.Father && !person.Mother) || (!person.Father && person.Mother);
-                }
-                return val;
-            }
-
-            let deathDate = CC7.ymdFix(mPerson.DeathDate);
-            if (deathDate == "") {
-                if (mPerson.deathDateDecade) {
-                    deathDate = mPerson.DeathDateDecade;
-                }
-            }
-            let birthLocation = mPerson.BirthLocation;
-            let birthLocationReversed = "";
-            if (birthLocation == null || typeof birthLocation == "undefined") {
-                birthLocation = "";
-            } else {
-                const bLocation2ways = CC7.location2ways(birthLocation);
-                birthLocation = bLocation2ways[0];
-                birthLocationReversed = bLocation2ways[1];
-            }
-            let deathLocation = mPerson.DeathLocation;
-            let deathLocationReversed = "";
-            if (deathLocation == null || typeof deathLocation == "undefined") {
-                deathLocation = "";
-            } else {
-                const dLocation2ways = CC7.location2ways(deathLocation);
-                deathLocation = dLocation2ways[0];
-                deathLocationReversed = dLocation2ways[1];
-            }
-
-            const privacyLevel = +mPerson.Privacy;
-            const privacyDetail = CC7.PRIVACY_LEVELS.get(privacyLevel);
-            let privacyImg = null;
-            let privacyTitle = "";
-            let dprivacy = "";
-            if (privacyDetail) {
-                privacyTitle = privacyDetail.title;
-                privacyImg = privacyDetail.img;
-                dprivacy = "data-privacy='" + privacyLevel + "'";
-            }
-
-            let firstName = mPerson.FirstName;
-            if (mPerson.MiddleName) {
-                firstName = mPerson.FirstName + " " + mPerson.MiddleName;
-            }
-
-            if (!mPerson.FirstName && mPerson.RealName) {
-                firstName = mPerson.RealName;
-            }
-
-            if (birthDate == "unknown") {
-                birthDate = "";
-            }
-            if (deathDate == "unknown") {
-                deathDate = "";
-            }
-
-            let dBirthDate;
-            if (mPerson.BirthDate) {
-                dBirthDate = mPerson.BirthDate.replaceAll("-", "");
-            } else if (mPerson.BirthDateDecade) {
-                dBirthDate = CC7.getApproxDate2(mPerson.BirthDateDecade).Date.replace("-", "").padEnd(8, "0");
-            } else {
-                dBirthDate = "00000000";
-            }
-
-            let dDeathDate;
-            if (mPerson.DeathDate) {
-                dDeathDate = mPerson.DeathDate.replaceAll("-", "");
-            } else if (mPerson.DeathDateDecade) {
-                dDeathDate = CC7.getApproxDate2(mPerson.DeathDateDecade).Date.replace("-", "").padEnd(8, "0");
-            } else {
-                dDeathDate = "00000000";
-            }
-
-            if (firstName == undefined) {
-                firstName = "Private";
-                mPerson.LastNameCurrent = "";
-                if (mPerson.Name.match(/Private/) == null) {
-                    mPerson.LastNameAtBirth = mPerson.Name.split("-")[1];
-                } else {
-                    mPerson.LastNameAtBirth = "Private";
-                }
-            }
-
-            const oLink = CC7.profileLink(mPerson.Name, firstName);
-            let managerLink;
-            let dManager;
-            if (mPerson.Manager) {
-                const mgrWtId = mPerson.Managers?.find((m) => m.Id == mPerson.Manager)?.Name || mPerson.Manager;
-                dManager = CC7.htmlEntities(mgrWtId);
-                managerLink = CC7.profileLink(dManager, dManager);
-            } else {
-                managerLink = mPerson.Name.startsWith("Private") ? "Unknown" : "Orphaned";
-                dManager = managerLink;
-            }
-
-            let degreeCell = "";
-            let touched = "";
-            let created = "";
-            let ddegree = "";
-            let dtouched = "";
-            let dcreated = "";
-            let ageAtDeathCell = "";
-            let dAgeAtDeath = "";
-            let diedYoung = false;
-            let diedVeryYoung = false;
-            let diedYoungIcon = "";
-            let diedYoungTitle = "";
-
-            let relNums = {
-                Parent_data: "",
-                Sibling_data: "",
-                Spouse_data: "",
-                Child_data: "",
-                Parent_cell: "",
-                Sibling_cell: "",
-                Spouse_cell: "",
-                Child_cell: "",
-            };
-
-            if ($("#cc7Container").length) {
-                degreeCell = "<td class='degree'>" + mPerson.Meta.Degrees + "°</td>";
-                ddegree = "data-degree='" + mPerson.Meta.Degrees + "'";
-                if (mPerson.Created) {
-                    created =
-                        "<td class='created aDate'>" +
-                        mPerson.Created.replace(/([0-9]{4})([0-9]{2})([0-9]{2}).*/, "$1-$2-$3") +
-                        "</td>";
-                    dcreated = "data-created='" + mPerson.Created + "'";
-                } else {
-                    created = "<td class='created aDate'></td>";
-                }
-
-                let mAgeAtDeath = CC7.ageAtDeath(mPerson);
-                let mAgeAtDeathNum = CC7.ageAtDeath(mPerson, false);
-
-                if (mAgeAtDeath === false && mAgeAtDeath !== "0") {
-                    mAgeAtDeath = "";
-                }
-                if (mAgeAtDeathNum < 0) {
-                    mAgeAtDeath = 0;
-                }
-                if (mAgeAtDeathNum < 5 && (mAgeAtDeath != false || mAgeAtDeathNum === 0)) {
-                    diedYoung = true;
-                    diedVeryYoung = true;
-                    diedYoungIcon = CC7.currentSettings["icons_options_veryYoung"];
-                    diedYoungTitle = "Died before age 5";
-                } else if (mAgeAtDeathNum < 16 && mAgeAtDeath != false) {
-                    diedYoung = true;
-                    diedYoungIcon = CC7.currentSettings["icons_options_young"];
-                    diedYoungTitle = "Died before age 16";
-                }
-
-                ageAtDeathCell = "<td class='age-at-death'>" + mAgeAtDeath + "</td>";
-                dAgeAtDeath = "data-age-at-death='" + mAgeAtDeathNum + "'";
-
-                if (mPerson.Touched) {
-                    touched =
-                        "<td class='touched aDate'>" +
-                        CC7.changesLink(
-                            mPerson.Name,
-                            mPerson.Touched.replace(/([0-9]{4})([0-9]{2})([0-9]{2}).*/, "$1-$2-$3")
-                        ) +
-                        "</td>";
-                    dtouched = "data-touched='" + mPerson.Touched + "'";
-                } else {
-                    touched = "<td class='touched aDate'></td>";
-                    dtouched = "data-touched=''";
-                }
-
-                relNums = {};
-                const rArr = ["Sibling", "Spouse", "Child"];
-                rArr.forEach(function (aR) {
-                    let cellClass = "class='number'";
-                    // For spouse count we check the Spouses arrays since it contains everyone, not just
-                    // those present in the returned data
-                    const realAr = aR == "Spouse" ? "Spouses" : aR;
-                    if (mPerson[realAr].length) {
-                        relNums[aR] = mPerson[realAr].length;
-                    } else {
-                        relNums[aR] = "";
-                    }
-                    relNums[aR + "_data"] = "data-" + aR + "='" + relNums[aR] + "'";
-                    let word = aR + "s";
-                    if (aR == "Child") {
-                        word = "Children";
-                        if (diedYoung && relNums[aR] == "") {
-                            const imgClass = diedVeryYoung ? "diedVeryYoungImg" : "diedYoungImg";
-                            relNums[aR] = `<img class="${imgClass}" src="${CC7.imagePath(diedYoungIcon)}" />`;
-                            cellClass = "class='number diedYoung'";
-                            word = diedYoungTitle;
-                        } else if (mPerson.NoChildren == 1) {
-                            cellClass = "class='none number'";
-                        }
-                    } else if (
-                        aR == "Sibling" &&
-                        mPerson.Parent.length == 2 &&
-                        mPerson.Parent[0].NoChildren &&
-                        mPerson.Parent[0].NoChildren
-                    ) {
-                        cellClass = "class='none number'";
-                    }
-                    if (aR == "Spouse") {
-                        if (mPerson.DataStatus?.Spouse == "blank" || (diedYoung && relNums[aR] == "")) {
-                            cellClass = "class='none number'";
-                        }
-                    }
-                    relNums[aR + "_cell"] = "<td " + cellClass + " title='" + word + "'>" + relNums[aR] + "</td>";
-                });
-
-                if (!mPerson.Father && !mPerson.Mother) {
-                    relNums["Parent"] = 0;
-                    relNums["Parent_data"] = "data-Parent='0'";
-                    relNums["Parent_cell"] = "<td class='noParents number' title='Missing parents'>0</td>";
-                } else if (!mPerson.Father) {
-                    relNums["Parent"] = 1;
-                    relNums["Parent_data"] = "data-Parent='1'";
-                    relNums["Parent_cell"] = "<td class='noFather number' title='Missing father'>1</td>";
-                } else if (!mPerson.Mother) {
-                    relNums["Parent"] = 1;
-                    relNums["Parent_data"] = "data-Parent='1'";
-                    relNums["Parent_cell"] = "<td class='noMother number' title='Missing mother'>1</td>";
-                } else {
-                    relNums["Parent"] = 2;
-                    relNums["Parent_data"] = "data-Parent='2'";
-                    relNums["Parent_cell"] = "<td class='number' title='Parents'>2</td>";
-                }
-            }
-
-            let gender = mPerson.Gender;
-            if (mPerson?.DataStatus?.Gender == "blank") {
-                gender = "blank";
-            }
-
-            const aLine = $(
-                "<tr " +
-                    dprivacy +
-                    " " +
-                    ddegree +
-                    " " +
-                    dAgeAtDeath +
-                    " " +
-                    dtouched +
-                    " " +
-                    dcreated +
-                    " " +
-                    relNums["Parent_data"] +
-                    " " +
-                    relNums["Sibling_data"] +
-                    " " +
-                    relNums["Spouse_data"] +
-                    " " +
-                    relNums["Child_data"] +
-                    " data-id='" +
-                    mPerson.Id +
-                    "' data-name='" +
-                    CC7.htmlEntities(mPerson.Name) +
-                    "' data-firstname='" +
-                    CC7.htmlEntities(firstName) +
-                    "' data-lnab='" +
-                    CC7.htmlEntities(mPerson.LastNameAtBirth) +
-                    "'  data-lnc='" +
-                    CC7.htmlEntities(mPerson.LastNameCurrent) +
-                    "' data-birthdate='" +
-                    dBirthDate +
-                    "' data-deathdate='" +
-                    dDeathDate +
-                    "' data-birthlocation='" +
-                    CC7.htmlEntities(birthLocation) +
-                    "' data-birthlocation-reversed='" +
-                    CC7.htmlEntities(birthLocationReversed) +
-                    "' data-deathlocation='" +
-                    CC7.htmlEntities(deathLocation) +
-                    "' data-deathlocation-reversed='" +
-                    CC7.htmlEntities(deathLocationReversed) +
-                    "' data-manager='" +
-                    dManager +
-                    "' class='" +
-                    gender +
-                    `'><td ${
-                        mPerson.hasBioIssues
-                            ? CC7.mustHighlight(mPerson.bioCheckReport)
-                                ? "class='privBio bioIssue' title='Click to see Bio Check Report'"
-                                : "class='privBio bioIssue2' title='Click to see Bio Check Report'"
-                            : "class='privBio'"
-                    }>` +
-                    (privacyImg
-                        ? "<img class='privacyImage' src='" + privacyImg + "' title='" + privacyTitle + "'>"
-                        : "") +
-                    `</td><td><img class='familyHome' src='./views/cc7/images/Home_icon.png' title="Click to see ${firstName}'s family sheet"></td>` +
-                    `<td><img class='timelineButton' src='./views/cc7/images/timeline.png' title="Click to see a timeline for ${firstName}"></td>` +
-                    degreeCell +
-                    relNums["Parent_cell"] +
-                    relNums["Sibling_cell"] +
-                    relNums["Spouse_cell"] +
-                    relNums["Child_cell"] +
-                    "<td class='connectionsName' >" +
-                    oLink +
-                    "</td><td class='lnab'>" +
-                    (mPerson.LastNameAtBirth.startsWith("Private")
-                        ? mPerson.LastNameAtBirth
-                        : "<a target='_blank' href='https://www.wikitree.com/index.php?title=Special:Surname&order=name&layout=table&s=" +
-                          CC7.htmlEntities(mPerson.LastNameAtBirth) +
-                          "'>" +
-                          mPerson.LastNameAtBirth +
-                          "</a>") +
-                    "</td><td class='lnc'><a   target='_blank' href='https://www.wikitree.com/index.php?title=Special:Surname&order=name&layout=table&s=" +
-                    CC7.htmlEntities(mPerson.LastNameCurrent) +
-                    "'>" +
-                    mPerson.LastNameCurrent +
-                    "</a></td><td class='aDate birthdate'>" +
-                    birthDate +
-                    "</td><td class='location birthlocation'>" +
-                    CC7.htmlEntities(birthLocation) +
-                    "</td><td  class='aDate deathdate'>" +
-                    deathDate +
-                    "</td><td class='location deathlocation'>" +
-                    CC7.htmlEntities(deathLocation) +
-                    "</td>" +
-                    ageAtDeathCell +
-                    "<td class='managerCell'>" +
-                    managerLink +
-                    "</td>" +
-                    created +
-                    touched +
-                    "</tr>"
-            );
-
-            aTable.find("tbody").append(aLine);
-        }
-
-        if (CC7.getCookie("w_wideTable") == "0") {
-            CC7.setOverflow("visible");
-        } else {
-            CC7.setOverflow("auto");
-        }
-
-        // Provide a way to examine the data record of a specific person
-        $("img.privacyImage, .bioIssue")
-            .off("click")
-            .on("click", function (event) {
-                event.stopImmediatePropagation();
-                const id = $(this).closest("tr").attr("data-id");
-                const p = window.people.get(+id);
-                if (event.altKey) {
-                    console.log(`${p.Name}, ${p.BirthNamePrivate}`, p);
-                } else if (p.hasBioIssues) {
-                    CC7.showBioCheckReport($(this));
-                }
-            });
-        $("img.familyHome")
-            .off("click")
-            .on("click", function () {
-                CC7.showFamilySheet($(this));
-            });
-        $("img.timelineButton")
-            .off("click")
-            .on("click", function (event) {
-                CC7.showTimeline($(this));
-            });
-
-        aTable.find("th[id]").each(function () {
-            CC7.sortByThis($(this));
-        });
-        CC7.addWideTableButton();
-        if ($("#hierarchyViewButton").length == 0) {
-            $("#wideTableButton").before(
-                $(
-                    "<select id='cc7Subset' title='Select which profiles should be displayed'>" +
-                        "<option value='all' selected>All</option>" +
-                        "<option value='ancestors' title='Direct ancestors only'>Ancestors</option>" +
-                        "<option value='descendants' title='Direct descendants only'>Descendants</option>" +
-                        '<option value="above" title="Anyone that can be reached by first following a parent link">All "Above"</option>' +
-                        '<option value="below" title="Anyone that can be reached by first following a non-parent link">All "Below"</option>' +
-                        '<option value="missing-links" title="People that may be missing family members" link">Missing Family</option>' +
-                        "</select>" +
-                        "<button class='button small viewButton' id='hierarchyViewButton'>Hierarchy</button>" +
-                        "<button class='button small viewButton' id='listViewButton'>List</button>" +
-                        "<button class='button small viewButton active' id='tableViewButton'>Table</button>"
-                )
-            );
-        }
-        $("#cc7Subset")
-            .off("change")
-            .on("change", function () {
-                const curTableId = $("table.active").attr("id");
-                if (curTableId == "lanceTable") {
-                    CC7.lanceView();
-                } else if (curTableId == "peopleTable") {
-                    drawPeopleTable();
-                }
-                if ($("#cc7Subset").val() == "missing-links") {
-                    CC7.showMissingLinksButtons();
-                } else {
-                    $("#mlButtons").hide();
-                }
-            });
-
-        function drawPeopleTable() {
-            const subset = $("#cc7Subset").val();
-            const caption = CC7.tableCaption(window.rootId, $("#cc7Degree").val(), !CC7.ONE_DEGREE); // cc7Subset is disabled for single degrees
-            let subsetWord = "";
-            switch (subset) {
-                case "ancestors":
-                    subsetWord = " (Ancestors Only)";
-                    break;
-                case "descendants":
-                    subsetWord = " (Descendants Only)";
-                    break;
-                case "above":
-                    subsetWord = ' ("Above" Only)';
-                    break;
-                case "below":
-                    subsetWord = ' ("Below" Only)';
-                    break;
-                case "missing-links":
-                    subsetWord = " (Missing Family)";
-                    break;
-                default:
-                    break;
-            }
-            CC7.addPeopleTable(caption + subsetWord);
-        }
-
-        $("#listViewButton")
-            .off("click")
-            .on("click", function () {
-                $(".viewButton").removeClass("active");
-                $(this).addClass("active");
-                $("#peopleTable,#hierarchyView").hide();
-                if ($("#lanceTable").length == 0 || !$("#lanceTable").hasClass($("#cc7Subset").val())) {
-                    CC7.lanceView();
-                } else {
-                    $("#lanceTable").show().addClass("active");
-                    $("#wideTableButton").hide();
-                }
-                $("#cc7Subset").show();
-                if ($("#cc7Subset").val() == "missing-links") {
-                    CC7.showMissingLinksButtons();
-                }
-            });
-        $("#hierarchyViewButton")
-            .off("click")
-            .on("click", function () {
-                $(".viewButton").removeClass("active");
-                $(this).addClass("active");
-                $("#peopleTable,#lanceTable").hide().removeClass("active");
-                if ($("#hierarchyView").length == 0) {
-                    CC7.showShakingTree(function () {
-                        // We only call hierarchyCC7 after a timeout in order to give the shaking tree
-                        // animation a change to complete first
-                        setTimeout(() => CC7.hierarchyCC7(), 10);
-                    });
-                    $("#wideTableButton").hide();
-                } else {
-                    $("#hierarchyView").show();
-                }
-                $("#cc7Subset").hide();
-                $("#mlButtons").hide();
-            });
-        $("#tableViewButton")
-            .off("click")
-            .on("click", function () {
-                $(".viewButton").removeClass("active");
-                $(this).addClass("active");
-                $("#hierarchyView,#lanceTable").hide().removeClass("active");
-                $("#cc7Subset").show();
-                if ($("#peopleTable").hasClass($("#cc7Subset").val())) {
-                    // We don't have to re-draw he table
-                    $("#peopleTable").show().addClass("active");
-                    $("#wideTableButton").show();
-                } else {
-                    drawPeopleTable();
-                }
-                if ($("#cc7Subset").val() == "missing-links") {
-                    CC7.showMissingLinksButtons();
-                }
-            });
-
-        CC7.hideShakingTree();
-
-        CC7.addFiltersToPeopleTable();
-        aTable.css({
-            "overflow-x": "auto",
-            "width": "100%",
-        });
-
-        if ($("#cc7excel").length == 0) {
-            $(
-                '<button id="cc7excel" title="Export an Excel file." class="small button" style="display: inline-block;">Excel</button>'
-            ).insertAfter($("#loadButton"));
-            $("#cc7excel")
-                .off("click")
-                .on("click", function () {
-                    CC7.cc7excelOut(CC7.EXCEL);
-                });
-        }
-        if ($("#cc7csv").length == 0) {
-            $(
-                '<button id="cc7csv" title="Export a CSV file." class="small button" style="display: inline-block;">CSV</button>'
-            ).insertAfter($("#loadButton"));
-            $("#cc7csv")
-                .off("click")
-                .on("click", function () {
-                    CC7.cc7excelOut(CC7.CSV);
-                });
-        }
-        $("#lanceTable").removeClass("active");
-        aTable.addClass("active");
-        aTable.floatingScroll();
-    }
-
-    static addFiltersToPeopleTable() {
-        const table = document.querySelector("#peopleTable");
-        const hasTbody = table.querySelector("tbody") !== null;
-        const hasThead = table.querySelector("thead") !== null;
-        const headerRow = hasThead
-            ? table.querySelector("thead tr:first-child")
-            : hasTbody
-            ? table.querySelector("tbody tr:first-child")
-            : table.querySelector("tr:first-child");
-
-        let headerCells = headerRow.querySelectorAll("th");
-        const originalHeaderCells = headerCells;
-        const isFirstRowHeader = headerCells.length > 0;
-        if (!isFirstRowHeader) {
-            const firstRowCells = headerRow.querySelectorAll("td");
-            const dummyHeaderRow = document.createElement("tr");
-            firstRowCells.forEach(() => {
-                const emptyHeaderCell = document.createElement("th");
-                dummyHeaderRow.appendChild(emptyHeaderCell);
-            });
-            headerRow.parentElement.insertBefore(dummyHeaderRow, headerRow);
-            headerCells = dummyHeaderRow.querySelectorAll("th");
-        }
-
-        const filterRow = document.createElement("tr");
-        filterRow.classList.add("cc7filter-row");
-
-        headerCells.forEach((headerCell, i) => {
-            const filterCell = document.createElement("th");
-            if (i == 0) {
-                filterCell.title = "Select a column value to which to limit the rows";
-                filterRow.appendChild(filterCell);
-                // $(
-                // "<select id='cc7PBFilter' title='Select which Privacy/BioCheck rows should be displayed'>"
-                // <option value='all' title="Everything" selected>&nbsp;</option>"
-                // <option value='bioBad' title="With Bio Check Issues">./views/cc7/images/checkbox-cross-red-icon.png</option>"
-                // <option value='bioOK' title="No Bio Check Issues">./views/cc7/images/checkmark-box-green-icon.png</option>"
-                // <option value='60' title='Privacy: Open'>./views/cc7/images/privacy_open.png</option>" +
-                // <option value='50' title='Public'>./views/cc7/images/privacy_public.png</option>" +
-                // <option value='40' title='Private with Public Bio and Tree'>./views/cc7/images/privacy_public-tree.png</option>"
-                // etc.
-                // );
-                const filterSelect = document.createElement("select");
-                filterSelect.id = "cc7PBFilter";
-                filterSelect.title = "Select which Privacy/BioCheck rows should be displayed";
-                let filterOption = document.createElement("option");
-                filterOption.value = "all";
-                filterOption.title = "Everything";
-                filterOption.text = " ";
-                filterSelect.appendChild(filterOption);
-                if (CC7.currentSettings["biocheck_options_biocheckOn"]) {
-                    filterOption = document.createElement("option");
-                    filterOption.value = "bioBad";
-                    filterOption.title = "With Bio Check Issues";
-                    filterOption.text = "./views/cc7/images/checkbox-cross-red-icon.png";
-                    filterSelect.appendChild(filterOption);
-                    filterOption = document.createElement("option");
-                    filterOption.value = "bioOK";
-                    filterOption.title = "No Bio Check Issue";
-                    filterOption.text = "./views/cc7/images/checkmark-box-green-icon.png";
-                    filterSelect.appendChild(filterOption);
-                }
-                for (const privLevel of CC7.PRIVACY_LEVELS.keys()) {
-                    const privacy = CC7.PRIVACY_LEVELS.get(privLevel);
-                    filterOption = document.createElement("option");
-                    filterOption.value = privLevel;
-                    filterOption.title = privacy.title;
-                    filterOption.text = privacy.img;
-                    filterSelect.appendChild(filterOption);
-                }
-                filterCell.appendChild(filterSelect);
-                return;
-            } else if (i == 1) {
-                filterCell.colSpan = 2;
-                filterCell.style.textAlign = "right";
-                filterCell.innerHTML = "Filters";
-                filterCell.title =
-                    "Show only rows with these column values. > and < may be used for numerical columns.";
-                filterRow.appendChild(filterCell);
-                return;
-            }
-            const headerCellText = headerCell.textContent.trim();
-            const originalHeaderCellText = originalHeaderCells[i].textContent.trim();
-            if (!["Pos.", "P/B"].includes(headerCellText) && !["Pos.", "P/B", ""].includes(originalHeaderCellText)) {
-                // console.log(headerCellText);
-                filterCell.title = "Enter a column value to which to limit the rows";
-                const filterInput = document.createElement("input");
-                filterInput.type = "text";
-                filterInput.classList.add("filter-input");
-
-                // Check the length of the text in the first ten cells of the column
-                const rows = hasTbody ? table.querySelectorAll("tbody tr") : table.querySelectorAll("tr");
-                let isNumeric = 0;
-                let isDate = 0;
-                let isText = 0;
-                for (let j = 1; j < Math.min(50, rows.length); j++) {
-                    if (rows[j]) {
-                        const cellText = rows[j].children[i].textContent.trim();
-                        const cellTextStripped = cellText.replace(/[<>~]?(\d+)°?/g, "$1");
-                        const dateMatch = cellText.match(/(\d{4})(-(\d{2})-(\d{2}))?/);
-                        if (dateMatch) {
-                            isDate++;
-                        } else if (CC7.isNumeric(cellTextStripped)) {
-                            isNumeric++;
-                        } else if (cellText !== "") {
-                            isText++;
-                        }
-                    }
-                }
-
-                let maxVal;
-                if (isNumeric > isDate && isNumeric > isText) {
-                    maxVal = "isNumeric";
-                } else if (isDate > isNumeric && isDate > isText) {
-                    maxVal = "isDate";
-                } else {
-                    maxVal = "isText";
-                }
-
-                if (maxVal == "isNumeric") {
-                    filterInput.classList.add("numeric-input");
-                } else if (maxVal == "isDate") {
-                    filterInput.classList.add("date-input");
-                } else {
-                    filterInput.classList.add("text-input");
-                    filterInput.addEventListener("input", CC7.stripLtGt);
-                }
-
-                filterCell.appendChild(filterInput);
-            }
-            if (i > 2) {
-                filterRow.appendChild(filterCell);
-            }
-        });
-
-        if (isFirstRowHeader) {
-            headerRow.parentElement.insertBefore(filterRow, headerRow.nextSibling);
-        } else {
-            headerRow.parentElement.insertBefore(filterRow, headerRow);
-        }
-
-        function formatOptions(option) {
-            // option:
-            // {
-            //     "id": "value attribute" || "option text",
-            //     "text": "label attribute" || "option text",
-            //     "element": HTMLOptionElement
-            // }
-            if (!option.id || option.id == "all") {
-                return option.text;
-            }
-            return $(`<img class="privacyImage" src="./${option.text}"/>`);
-        }
-        $("#cc7PBFilter").select2({
-            templateResult: formatOptions,
-            templateResult: formatOptions,
-            templateSelection: formatOptions,
-            dropdownParent: $("#cc7Container"),
-            minimumResultsForSearch: Infinity,
-            width: "2em",
-        });
-
-        document.querySelectorAll(".filter-input").forEach((input) => {
-            input.addEventListener("input", CC7.filterListener);
-        });
-        $("#cc7PBFilter").off("select2:select").on("select2:select", CC7.filterListener);
-
-        // Add Clear Filters button
-        $("#clearTableFiltersButton").remove();
-        const clearFiltersButton = document.createElement("button");
-        clearFiltersButton.textContent = "X";
-        clearFiltersButton.title = "Clear Filters";
-        clearFiltersButton.id = "clearTableFiltersButton";
-        //  clearFiltersButton.style.position = "absolute";
-        clearFiltersButton.addEventListener("click", CC7.clearFilterClickListener);
-
-        $(clearFiltersButton).insertAfter($("#wideTableButton"));
-        clearFiltersButton.textContent = "Clear Filters";
-
-        // Initially hide the button
-        clearFiltersButton.style.display = "none";
-    }
-
-    static clearFilterClickListener() {
-        document.querySelectorAll(".filter-input").forEach((input) => {
-            input.value = "";
-        });
-        $("#cc7PBFilter").val("all").trigger("change");
-        CC7.filterFunction();
-        CC7.updateClearFiltersButtonVisibility();
-    }
-
-    static filterFunction() {
-        const table = document.querySelector("#peopleTable");
-        const hasTbody = table.querySelector("tbody") !== null;
-        const hasThead = table.querySelector("thead") !== null;
-        const rows = hasTbody ? table.querySelectorAll("tbody tr") : table.querySelectorAll("tr");
-        const filterInputs = table.querySelectorAll(".filter-input");
-
-        rows.forEach((row, rowIndex) => {
-            // Skip first row only if there's no 'thead'
-            if (!hasThead && rowIndex === 0) {
-                return;
-            }
-
-            // Skip if row is a filter-row or contains 'th' elements
-            if (row.classList.contains("cc7filter-row") || row.querySelector("th")) {
-                return;
-            }
-
-            const filters = [];
-            filterInputs.forEach((input, inputIndex) => {
-                let filterText = input.value.toLowerCase();
-                if (filterText.length == 0) {
-                    return;
-                }
-                const columnIndex =
-                    Array.from(input.parentElement.parentElement.children).indexOf(input.parentElement) + 1;
-                filters.push({
-                    input: input,
-                    cellText: row.children[columnIndex].textContent.toLowerCase(),
-                });
-            });
-
-            let displayRow = true;
-            filters.forEach((filter, inputIndex) => {
-                const input = filter.input;
-                let filterText = input.value.toLowerCase();
-                let cellText = filter.cellText;
-                const isDateColumn = input.classList.contains("date-input");
-                const isNumericColumn = input.classList.contains("numeric-input");
-                let operator;
-                if (["<", ">", "!", "?"].includes(filterText[0])) {
-                    operator = filterText[0];
-                    if (operator == "!") operator = "!=";
-                    filterText = filterText.slice(1);
-                    if (filterText.length == 0 && operator != "?") {
-                        return;
-                    }
-                }
-
-                // Perform the appropriate checks
-                // Note, since we want the && of all the filters, the code below should only set
-                // displayRow to false as and when necessary (and never set it to true).
-                if (operator == "?") {
-                    // Select rows with an empty cell in this column
-                    if (cellText != "") displayRow = false;
-                } else if (
-                    (isNumericColumn && filterText != "~") ||
-                    (isDateColumn && (operator == ">" || operator == "<"))
-                ) {
-                    // Use the operator to do an actual comparison
-                    if (filterText.length > 0) {
-                        if (operator) {
-                            filterText = parseFloat(filterText);
-                        } else if (!operator && isNumericColumn) {
-                            operator = "=="; // Default to equality if there's no operator
-                            filterText = parseFloat(filterText.replace(/[<>~]/g, ""));
-                        } else {
-                            filterText = `"${filterText}"`;
-                        }
-                        if (isDateColumn) {
-                            let year = cellText.slice(0, 4); // Get the year part of the date
-                            if (year.endsWith("s")) {
-                                year = year.slice(0, -1); // Remove the 's' for decade dates
-                            }
-                            cellText = parseFloat(year);
-                        } else if (isNumericColumn) {
-                            cellText = parseFloat(cellText.replace(/[<>~]/g, ""));
-                        } else {
-                            cellText = `"${cellText}"`;
-                        }
-                        if (!eval(cellText + operator + filterText)) {
-                            displayRow = false;
-                        }
-                    }
-                } else {
-                    // Perform partial matching and lip the result for the not (!) operator
-                    let aMatch = cellText.includes(filterText);
-                    if (operator == "!=") {
-                        aMatch = !aMatch;
-                    }
-                    if (!aMatch) displayRow = false;
-                }
-            });
-            // Add the Privacy/BioCheck filter
-            const pbFilterSelected = $("#cc7PBFilter").select2("data")[0];
-            const reqPrivacy = pbFilterSelected.id;
-            if (reqPrivacy != "all") {
-                if (reqPrivacy == "bioOK") {
-                    if (row.children[0].classList.contains("bioIssue")) displayRow = false;
-                } else if (reqPrivacy == "bioBad") {
-                    if (!row.children[0].classList.contains("bioIssue")) displayRow = false;
-                } else {
-                    const rowPrivacy = row.getAttribute("data-privacy");
-                    if (rowPrivacy != reqPrivacy) displayRow = false;
-                }
-            }
-
-            row.style.display = displayRow ? "" : "none";
-        });
-        $("#peopleTable").floatingScroll("update");
-    }
-
-    static anyFilterActive() {
-        return (
-            Array.from(document.querySelectorAll(".filter-input")).some((input) => input.value.trim() !== "") ||
-            $("#cc7PBFilter").select2("data")[0].id != "all"
-        );
-    }
-
-    static updateClearFiltersButtonVisibility() {
-        // let anyFilterHasText = Array.from(document.querySelectorAll(".filter-input")).some(
-        //     (input) => input.value.trim() !== ""
-        // );
-        // if ($("#cc7PBFilter").select2("data")[0].id != "all") anyFilterHasText = true;
-        const clearFiltersButton = document.querySelector("#clearTableFiltersButton");
-        clearFiltersButton.style.display = CC7.anyFilterActive() ? "inline-block" : "none";
-    }
-
-    static stripLtGt() {
-        if (this.value == ">" || this.value == "<") {
-            this.value = "";
-        }
-    }
-
-    static filterListener() {
-        CC7.filterFunction();
-        CC7.updateClearFiltersButtonVisibility();
-    }
-
-    static getApproxDate(theDate) {
-        let approx = false;
-        let aDate;
-        if (theDate.match(/0s$/) != null) {
-            aDate = theDate.replace(/0s/, "5");
-            approx = true;
-        } else {
-            const bits = theDate.split("-");
-            if (theDate.match(/00\-00$/) != null) {
-                aDate = bits[0] + "-07-02";
-                approx = true;
-            } else if (theDate.match(/-00$/) != null) {
-                aDate = bits[0] + "-" + bits[1] + "-" + "16";
-                approx = true;
-            } else {
-                aDate = theDate;
-            }
-        }
-        return { Date: aDate, Approx: approx };
-    }
-
-    static getApproxDate2(theDate) {
-        let approx = false;
-        let aDate;
-        if (theDate.match(/0s$/) != null) {
-            aDate = theDate.replace(/0s/, "5");
-            approx = true;
-        } else {
-            const bits = theDate.split("-");
-            if (theDate.match(/00\-00$/) != null || !bits[1]) {
-                aDate = bits[0] + "-07-02";
-                approx = true;
-            } else if (theDate.match(/-00$/) != null) {
-                aDate = bits[0] + "-" + bits[1] + "-" + "16";
-                approx = true;
-            } else {
-                aDate = theDate;
-            }
-        }
-        return { Date: aDate, Approx: approx };
-    }
-
-    static displayDates(fPerson) {
-        const mbdDatesStatus = CC7.bdDatesStatus(fPerson);
-        const bdStatus = mbdDatesStatus[0];
-        const ddStatus = mbdDatesStatus[1];
-
-        let fbd = "";
-        let fdd = "";
-
-        if (
-            fPerson["BirthDate"] != "" &&
-            fPerson["BirthDate"] != "0000-00-00" &&
-            typeof fPerson["BirthDate"] != "undefined" &&
-            fPerson["BirthDate"] != "unknown"
-        ) {
-            fbd = fPerson["BirthDate"].split("-")[0];
-        } else if (typeof fPerson["BirthDateDecade"] != "undefined" && fPerson["BirthDateDecade"] != "unknown") {
-            fbd = fPerson["BirthDateDecade"];
-        } else {
-            fbd = "";
-        }
-
-        if (typeof fPerson["IsLiving"] != "undefined") {
-            if (fPerson["IsLiving"] == 1) {
-                fdd = "living";
-            }
-        }
-        if (fdd == "") {
-            if (
-                fPerson["DeathDate"] != "" &&
-                fPerson["DeathDate"] != "0000-00-00" &&
-                typeof fPerson["DeathDate"] != "undefined"
-            ) {
-                fdd = fPerson["DeathDate"].split("-")[0];
-            } else if (typeof fPerson["DeathDateDecade"] != "undefined" && fPerson["DeathDateDecade"] != "unknown") {
-                fdd = fPerson["DeathDateDecade"];
-            } else {
-                fdd = "";
-            }
-        }
-
-        const fDates = "(" + bdStatus + fbd + " - " + ddStatus + fdd + ")";
-        return fDates;
-    }
-
-    static displayFullDates(fPerson, showStatus = true) {
-        const mbdDatesStatus = CC7.bdDatesStatus(fPerson);
-        const bdStatus = mbdDatesStatus[0];
-        const ddStatus = mbdDatesStatus[1];
-
-        let fbd = "";
-        let fdd = "";
-
-        const fDates = [];
-
-        if (
-            fPerson["BirthDate"] != "" &&
-            fPerson["BirthDate"] != "0000-00-00" &&
-            typeof fPerson["BirthDate"] != "undefined"
-        ) {
-            const fbds = fPerson["BirthDate"].split("-");
-            if (fbds[0] == "unkno5") {
-                fbd = "";
-            } else {
-                fbd = CC7.getDateFormat(fbds);
-            }
-        } else if (typeof fPerson["BirthDateDecade"] != "undefined") {
-            fbd = fPerson["BirthDateDecade"];
-        } else {
-            fbd = "";
-        }
-
-        if (fbd == "") {
-            fDates.push("");
-        } else if (showStatus == false) {
-            fDates.push(fbd);
-        } else {
-            fDates.push(bdStatus + fbd);
-        }
-
-        if (typeof fPerson["IsLiving"] != "undefined") {
-            if (fPerson["IsLiving"] == 1) {
-                fdd = "living";
-            }
-        }
-        if (fdd == "") {
-            if (
-                fPerson["DeathDate"] != "" &&
-                fPerson["DeathDate"] != "0000-00-00" &&
-                typeof fPerson["DeathDate"] != "undefined"
-            ) {
-                const fdds = fPerson["DeathDate"].split("-");
-                if (fdds[0] == "unkno5") {
-                    fdd = "";
-                } else {
-                    fdd = CC7.getDateFormat(fdds);
-                }
-            } else if (typeof fPerson["DeathDateDecade"] != "undefined") {
-                if (fPerson["DeathDateDecade"] != "unknown") {
-                    fdd = fPerson["DeathDateDecade"];
-                }
-            } else {
-                fdd = "";
-            }
-        }
-
-        if (fdd == "") {
-            fDates.push("");
-        } else if (showStatus == false) {
-            fDates.push(fdd);
-        } else {
-            fDates.push(ddStatus + fdd);
-        }
-
-        return fDates;
-    }
-
-    static sortList(aList, dataField, reverse = 0) {
-        const rows = aList.find("li");
-        rows.sort(function (a, b) {
-            if (reverse == 0) {
-                return $(a).data(dataField).localeCompare($(b).data(dataField));
-            } else {
-                return $(b).data(dataField).localeCompare($(a).data(dataField));
-            }
-        });
-        rows.appendTo(aList);
-        CC7.secondarySort3(aList, "lnab", "first-name", 1);
-    }
-
-    static formDisplayName(aPerson, aName) {
-        return aPerson.Name.startsWith("Private")
-            ? aPerson.LastNameAtBirth || "Private"
-            : aName.withParts(CC7.WANTED_NAME_PARTS);
-    }
-
-    static async lanceView() {
-        $("#peopleTable").hide();
-        const surnames = {
-            degree_1: [],
-            degree_2: [],
-            degree_3: [],
-            degree_4: [],
-            degree_5: [],
-            degree_6: [],
-            degree_7: [],
-        };
-        const subset = $("#cc7Subset").val();
-        const lanceTable = $(
-            `<table id='lanceTable' class="${subset}">` +
-                "<thead>" +
-                "<tr></tr>" +
-                "</thead>" +
-                "<tbody>" +
-                "<tr></tr>" +
-                "</tbody>" +
-                "</table>"
-        );
-        if ($("#lanceTable").length) {
-            $("#lanceTable").eq(0).replaceWith(lanceTable);
-        } else {
-            $("#peopleTable").before(lanceTable);
-        }
-        for (let i = 1; i < 8; i++) {
-            let aHeading = $("<th id='degreeHeading_" + i + "'><span>" + i + ".</span><ol></ol></th>");
-            lanceTable.find("thead tr").append(aHeading);
-            let aCell = $("<td id='degree_" + i + "'><ol></ol></td>");
-            lanceTable.find("tbody tr").append(aCell);
-        }
-
-        for (let aPerson of window.people.values()) {
-            if (aPerson.Hide) continue;
-            switch (subset) {
-                case "above":
-                    if (!aPerson.isAbove) continue;
-                    break;
-
-                case "below":
-                    if (aPerson.isAbove) continue;
-                    break;
-
-                case "ancestors":
-                    if (aPerson.isAncestor) break;
-                    continue;
-
-                case "descendants":
-                    if (typeof aPerson.isAncestor != "undefined" && !aPerson.isAncestor) break;
-                    continue;
-
-                default:
-                    break;
-            }
-            if (!aPerson.Missing) {
-                CC7.addMissingBits(aPerson);
-            }
-            const theDegree = aPerson.Meta.Degrees;
-            const aName = new PersonName(aPerson);
-            const theName = CC7.formDisplayName(aPerson, aName);
-            const theParts = aName.getParts(["LastNameAtBirth", "FirstName"]);
-            const theLNAB = theParts.get("LastNameAtBirth");
-            const theFirstName = theParts.get("FirstName");
-
-            if (CC7.isOK(theDegree) && theDegree <= window.cc7Degree) {
-                if (!surnames["degree_" + theDegree].includes(theLNAB)) {
-                    // Add the LNAB to the appropriate degree column header
-                    surnames["degree_" + theDegree].push(theLNAB);
-                    const anLi2 = $(
-                        "<li data-lnab='" +
-                            theLNAB +
-                            "' data-name='" +
-                            aPerson.Name +
-                            "'><a target='_blank' href='https://www.wikitree.com/genealogy/" +
-                            theLNAB +
-                            "'>" +
-                            theLNAB +
-                            "</a></li>"
-                    );
-                    $("#degreeHeading_" + theDegree)
-                        .find("ol")
-                        .append(anLi2);
-                    anLi2
-                        .find("a")
-                        .off("click")
-                        .on("click", function (e) {
-                            e.preventDefault();
-
-                            const degreeNum = $(this).closest("th").prop("id").slice(-1);
-                            const theList = $("#degree_" + degreeNum).find("li");
-                            if ($(this).attr("data-clicked") == 1) {
-                                theList.show();
-                                $("#lanceTable thead a").each(function () {
-                                    $(this).attr("data-clicked", 0);
-                                });
-                            } else {
-                                const thisLNAB = $(this).closest("li").data("lnab");
-                                $("#lanceTable thead a").each(function () {
-                                    $(this).attr("data-clicked", 0);
-                                });
-                                $(this).attr("data-clicked", 1);
-                                theList.each(function () {
-                                    if ($(this).data("lnab") != thisLNAB) {
-                                        $(this).hide();
-                                    } else {
-                                        $(this).show();
-                                    }
-                                });
-                            }
-                        });
-                }
-                // Add the person to the appropriate column
-                const linkName = CC7.htmlEntities(aPerson.Name);
-                const missing = CC7.missingThings(aPerson);
-                const anLi = $(
-                    "<li " +
-                        missing.missingBit +
-                        " data-first-name='" +
-                        theFirstName +
-                        "' data-lnab='" +
-                        theLNAB +
-                        "' data-name=\"" +
-                        aPerson.Name +
-                        '">' +
-                        CC7.profileLink(linkName, theName) +
-                        " " +
-                        missing.missingIcons +
-                        "</li>"
-                );
-                $("#degree_" + theDegree)
-                    .find("ol")
-                    .append(anLi);
-            }
-        }
-
-        $("td ol,th ol").each(function () {
-            CC7.sortList($(this), "lnab");
-        });
-        $("#peopleTable").removeClass("active");
-        $("#lanceTable").addClass("active");
-    }
-
-    static addMissingBits(aPerson) {
-        aPerson.Missing = [];
-        if (!aPerson.Father) {
-            aPerson.Missing.push("Father");
-        }
-        if (!aPerson.Mother) {
-            aPerson.Missing.push("Mother");
-        }
-
-        if (
-            (CC7.ageAtDeath(aPerson, false) > 15 ||
-                aPerson?.DeathDate == "0000-00-00" ||
-                aPerson?.BirthDate == "0000-00-00") &&
-            ((aPerson?.DataStatus.Spouse != "known" && aPerson?.DataStatus.Spouse != "blank") ||
-                aPerson.NoChildren != "1")
-        ) {
-            if (aPerson.Spouse.length == 0 && aPerson?.DataStatus?.Spouse != "blank") {
-                aPerson.Missing.push("Spouse");
-            }
-            if (aPerson.Child.length == 0 && aPerson.NoChildren != "1") {
-                aPerson.Missing.push("Children");
-            }
-        }
-    }
-
-    static missingThings(aPerson) {
-        let missingBit = "";
-        let missingIcons = "";
-        aPerson.Missing.forEach(function (relation) {
-            missingBit += "data-missing-" + relation + "='1' ";
-            if (relation == "Father") {
-                missingIcons +=
-                    "<img title='Missing father' class='missingFather missingIcon' src='./views/cc7/images/blue_bricks_small.jpg'>";
-            }
-            if (relation == "Mother") {
-                missingIcons +=
-                    "<img title='Missing mother' class='missingMother missingIcon' src='./views/cc7/images/pink_bricks_small.jpg'>";
-            }
-            if (relation == "Spouse") {
-                missingIcons +=
-                    "<img title='Possible missing spouse' class='missingSpouse missingIcon' src='./views/cc7/images/spouse_bricks_small.png'>";
-            }
-            if (relation == "Children") {
-                missingIcons +=
-                    "<img title='Possible missing children' class='missingChildren missingIcon' src='./views/cc7/images/baby_bricks_small.png'>";
-            }
-        });
-        return { missingBit: missingBit, missingIcons: missingIcons };
-    }
-
-    static addPeopleToHierarchy(degree) {
-        $("#hierarchyView li[data-degree='" + degree + "']").each(function () {
-            const id = $(this).data("id");
-            const thisLI = $(this);
-            const aPerson = window.people.get(+id);
-
-            if (aPerson) {
-                CC7.assignRelationshipsFor(aPerson);
-                const familyMembers = [].concat(aPerson.Parent, aPerson.Sibling, aPerson.Spouse, aPerson.Child);
-
-                familyMembers.forEach(function (aMember) {
-                    CC7.addMissingBits(aMember);
-
-                    if (thisLI.closest('li[data-name="' + aMember.Name + '"]').length == 0) {
-                        const theDegree = aMember.Meta.Degrees;
-                        if (theDegree > aPerson.Meta.Degrees) {
-                            const aName = new PersonName(aMember);
-                            const theName = CC7.formDisplayName(aMember, aName);
-                            const theParts = aName.getParts(["LastNameAtBirth", "FirstName"]);
-                            const theLNAB = theParts.get("LastNameAtBirth");
-                            const theFirstName = theParts.get("FirstName");
-
-                            let relation = aMember.Relation;
-                            if (relation == "Child") {
-                                relation = CC7.mapGender(aMember.Gender, "Son", "Daughter", "Child");
-                            } else if (relation == "Sibling") {
-                                relation = CC7.mapGender(aMember.Gender, "Brother", "Sister", "Sibling");
-                            } else if (relation == "Parent") {
-                                relation = CC7.mapGender(aMember.Gender, "Father", "Mother", "Parent");
-                            } else if (relation == "Spouse") {
-                                relation = CC7.mapGender(aMember.Gender, "Husband", "Wife", "Spouse");
-                            }
-
-                            const missing = CC7.missingThings(aMember);
-                            const missingBit = missing.missingBit;
-                            const missingIcons = missing.missingIcons;
-                            const linkName = CC7.htmlEntities(aMember.Name);
-                            const bdDates = CC7.displayDates(aMember);
-                            const anLi = $(
-                                `<li data-birth-date='${aMember.BirthDate}' data-father='${aMember.Father}' ` +
-                                    `data-mother='${aMember.Mother}' data-id='${aMember.Id}' data-relation='${relation}' ` +
-                                    `${missingBit} data-lnab='${theLNAB}' data-degree='${aMember.Meta.Degrees}' ` +
-                                    `data-name=\"${aMember.Name}\" data-first-name='${theFirstName}'>${aMember.Meta.Degrees}° ` +
-                                    `<span class='relation ${relation}'>${relation}</span>: ` +
-                                    CC7.profileLink(linkName, theName) +
-                                    ` <span class='birthDeathDates'>${bdDates}</span> ${missingIcons}<ul></ul></li>`
-                            );
-                            thisLI.children("ul").append(anLi);
-                        }
-                    }
-                });
-            }
-        });
-    }
-
-    static hierarchyCC7() {
-        window.visibleDegrees = 0;
-        const hierarchySection = $(
-            "<section id='hierarchyView'><menu><button class='button small' id='showAllDegrees'>Expand All</button>" +
-                "<button id='showOneMoreDegree' class='button small'>+</button>" +
-                "<button id='showOneFewerDegree' class='button small'>−</button>" +
-                "</menu><ul></ul></section>"
-        );
-        hierarchySection.insertBefore($("#peopleTable"));
-
-        const aPerson = window.people.get(window.rootId);
-        const aName = new PersonName(aPerson);
-        const theName = CC7.formDisplayName(aPerson, aName);
-        const theParts = aName.getParts(["LastNameAtBirth", "FirstName"]);
-        const theLNAB = theParts.get("LastNameAtBirth");
-        const theFirstName = theParts.get("FirstName");
-        const linkName = CC7.htmlEntities(aPerson.Name);
-        const anLi = $(
-            `<li data-lnab='${theLNAB}' data-id='${aPerson.Id}' data-degree='${aPerson.Meta.Degrees}' ` +
-                `data-name=\"${aPerson.Name}\" data-first-name='${theFirstName}'>${aPerson.Meta.Degrees}° ` +
-                CC7.profileLink(linkName, theName) +
-                "<ul></ul></li>"
-        );
-        hierarchySection.children("ul").append(anLi);
-        for (let i = 0; i < window.cc7Degree; i++) {
-            CC7.addPeopleToHierarchy(i);
-        }
-        $("#hierarchyView li").each(function () {
-            let aButton;
-            if ($(this).find("ul li").length) {
-                aButton = $("<button class='toggler'>+</button>");
-                $(this).prepend(aButton);
-                CC7.showMissingCounts(aButton);
-                aButton.on("click", function (e) {
-                    e.preventDefault();
-                    $(this).parent().children("ul").toggle();
-                    $(this).text($(this).text() == "+" ? "−" : "+");
-                    CC7.showMissingCounts($(this));
-                });
-            } else {
-                aButton = $("<button class='nonToggler'></button>");
-                $(this).prepend(aButton);
-            }
-        });
-        $("#showAllDegrees").on("click", function (e) {
-            e.preventDefault();
-            window.visibleDegrees = window.cc7Degree;
-            if ($(this).text() == "Expand All") {
-                $(this).text("Collapse All");
-                $("#hierarchyView ul ul").show();
-                $("#hierarchyView button.toggler").text("−");
-                $("span.countBit").hide();
-                $("span.nodeCount").hide();
-            } else {
-                window.visibleDegrees = 0;
-                $(this).text("Expand All");
-                $("#hierarchyView ul ul").hide();
-                $("#hierarchyView button.toggler").text("+");
-                $("span.countBit").show();
-                $("span.nodeCount").show();
-            }
-        });
-        $("#showOneMoreDegree").on("click", function (e) {
-            e.preventDefault();
-            if (window.visibleDegrees < window.cc7Degree) {
-                window.visibleDegrees++;
-                let j = window.visibleDegrees;
-                for (let i = 0; i < j + 1; i++) {
-                    $("li[data-degree='" + i + "']")
-                        .parent()
-                        .show();
-
-                    $("li[data-degree='" + (i - 1) + "']")
-                        .children("span.countBit")
-                        .hide();
-                    $("li[data-degree='" + (i - 1) + "']")
-                        .children("span.nodeCount")
-                        .hide();
-
-                    $("li[data-degree='" + (i - 1) + "']")
-                        .closest("li")
-                        .children("button.toggler")
-                        .text("−");
-                }
-            }
-        });
-        $("#showOneFewerDegree").on("click", function (e) {
-            e.preventDefault();
-            if (window.visibleDegrees > 0) {
-                window.visibleDegrees--;
-                let j = window.visibleDegrees;
-
-                for (let i = window.cc7Degree; i >= j + 1; i--) {
-                    $("li[data-degree='" + i + "']")
-                        .parent()
-                        .hide();
-                    $("li[data-degree='" + (i - 1) + "']")
-                        .children("span.countBit")
-                        .show();
-                    $("li[data-degree='" + (i - 1) + "']")
-                        .children("span.nodeCount")
-                        .show();
-                    $("li[data-degree='" + (i - 1) + "']")
-                        .closest("li")
-                        .children("button.toggler")
-                        .text("+");
-                }
-            }
-        });
-
-        // Sort the lists
-        $("#hierarchyView ul").each(function () {
-            const theUL = $(this);
-            const rows = $(this).children("li");
-            rows.sort((a, b) =>
-                $(b).data("birth-date").replaceAll(/\-/g, "") < $(a).data("birth-date").replaceAll(/\-/g, "") ? 1 : -1
-            );
-            rows.appendTo($(this));
-            if ($(this).children("li[data-relation='Husband'],li[data-relation='Wife']").length) {
-                $(this)
-                    .children("li[data-relation='Husband'],li[data-relation='Wife']")
-                    .each(function () {
-                        let spouseId = $(this).data("id");
-                        theUL
-                            .children("li[data-father='" + spouseId + "'],li[data-father='" + spouseId + "']")
-                            .eq(0)
-                            .before($(this));
-                    });
-            }
-            theUL.children("li[data-relation='Father']").prependTo(theUL);
-        });
-        $("#cc7Subset").hide();
-        CC7.hideShakingTree();
-    }
-
-    static showMissingCounts(jq) {
-        const thisLI = jq.parent();
-        if (thisLI.children("span.countBit,span.nodeCount").length) {
-            if (thisLI.children("ul:visible").length) {
-                thisLI.children("span.countBit").hide();
-                thisLI.children("span.nodeCount").hide();
-            } else {
-                thisLI.children("span.countBit").show();
-                thisLI.children("span.nodeCount").show();
-            }
-        } else {
-            const allCount = thisLI.find("li").length;
-
-            const missingFathers =
-                thisLI.find("ul img.missingFather").length -
-                thisLI.find("li[data-degree='7'] img.missingFather").length;
-            const missingMothers =
-                thisLI.find("ul img.missingMother").length -
-                thisLI.find("li[data-degree='7'] img.missingMother").length;
-            const missingSpouses =
-                thisLI.find("ul img.missingSpouse").length -
-                thisLI.find("li[data-degree='7'] img.missingSpouse").length;
-            const missingChildren =
-                thisLI.find("ul img.missingChildren").length -
-                thisLI.find("li[data-degree='7'] img.missingChildren").length;
-            const countBit = $("<span class='countBit'></span>");
-            if (allCount > 0 && thisLI.children(".nodeCount").length == 0) {
-                $(
-                    `<span class='nodeCount' title='Number of profiles hidden under ${thisLI.attr(
-                        "data-first-name"
-                    )}'>${allCount}</span>`
-                ).appendTo(thisLI);
-            }
-            if ((missingFathers || missingMothers || missingSpouses || missingChildren) && allCount > 0) {
-                countBit.appendTo(thisLI);
-                if (missingFathers) {
-                    countBit.append(
-                        $(
-                            "<span title='People with missing father'>" +
-                                missingFathers +
-                                "<img class='missingFatherCount missingCountIcon' src='./views/cc7/images/blue_bricks_small.jpg'></span>"
-                        )
-                    );
-                }
-                if (missingMothers) {
-                    countBit.append(
-                        $(
-                            "<span title='People with missing mother'>" +
-                                missingMothers +
-                                "<img class='missingMotherCount missingCountIcon' src='./views/cc7/images/pink_bricks_small.jpg'></span>"
-                        )
-                    );
-                }
-                if (missingSpouses) {
-                    countBit.append(
-                        $(
-                            "<span title='People with possible missing spouse'>" +
-                                missingSpouses +
-                                "<img class='missingSpouseCount missingCountIcon' src='./views/cc7/images/spouse_bricks_small.png'></span>"
-                        )
-                    );
-                }
-                if (missingChildren) {
-                    countBit.append(
-                        $(
-                            "<span title='People with posiible missing children'>" +
-                                missingChildren +
-                                "<img class='missingChildrenCount missingCountIcon' src='./views/cc7/images/baby_bricks_small.png'></span>"
-                        )
-                    );
-                }
-            }
-        }
-    }
-
-    static async getDegreeAction(event) {
+    static async getOneDegreeOnly(event) {
         wtViewRegistry.clearStatus();
         window.people = new Map();
         window.rootId = null;
@@ -4030,7 +464,7 @@ export class CC7 {
             $("#cancelLoad").show();
             CC7.cancelLoadController = new AbortController();
             CC7.clearDisplay();
-            CC7.showShakingTree();
+            Utils.showShakingTree();
             $("#cc7Container").addClass("degreeView");
             const theDegree = +$("#cc7Degree").val();
             const degreeCounts = []; // currently this is being ignored
@@ -4043,7 +477,7 @@ export class CC7 {
             );
             if (resultByKeyAtD == "aborted") {
                 wtViewRegistry.showWarning("Profile retrieval cancelled.");
-                CC7.hideShakingTree();
+                Utils.hideShakingTree();
             } else {
                 console.log(
                     `Retrieved ${countAtD} profiles at degrees ${theDegree - 1} to ${theDegree + 1} in ${
@@ -4057,12 +491,13 @@ export class CC7 {
                 }
 
                 if (countAtD == 0) {
-                    CC7.hideShakingTree();
+                    Utils.hideShakingTree();
                     return;
                 }
                 window.rootId = +resultByKeyAtD[wtId].Id;
+                window.cc7Degree = theDegree;
                 CC7.populateRelativeArrays();
-                CC7.hideShakingTree();
+                Utils.hideShakingTree();
                 $("#degreesTable").remove();
                 $("#ancReport").remove();
                 $("#cc7Container").append(
@@ -4073,8 +508,8 @@ export class CC7 {
                             "</table>"
                     )
                 );
-                CC7.buildDegreeTableData(degreeCounts, theDegree, theDegree);
-                CC7.addPeopleTable(CC7.tableCaption(wtId, theDegree, CC7.ONE_DEGREE));
+                CC7.buildDegreeTableData(degreeCounts, theDegree);
+                PeopleTable.addPeopleTable(PeopleTable.tableCaption());
                 $("#cc7Subset").prop("disabled", true);
             }
             $("#getPeopleButton").prop("disabled", false);
@@ -4153,14 +588,14 @@ export class CC7 {
         try {
             const result = await WikiTreeAPI.postToAPI(
                 {
-                    appId: CC7.APP_ID,
+                    appId: Settings.APP_ID,
                     action: "getPeople",
                     keys: key,
                     nuclear: degree + 1,
                     minGeneration: degree - 1,
                     start: start,
                     limit: limit,
-                    fields: CC7.currentSettings["biocheck_options_biocheckOn"]
+                    fields: Settings.current["biocheck_options_biocheckOn"]
                         ? CC7.GET_PEOPLE_FIELDS + ",Bio"
                         : CC7.GET_PEOPLE_FIELDS,
                 },
@@ -4212,7 +647,7 @@ export class CC7 {
                 person.Child = [];
                 person.Marriage = {};
 
-                if (CC7.currentSettings["biocheck_options_biocheckOn"]) {
+                if (Settings.current["biocheck_options_biocheckOn"]) {
                     const bioPerson = new BioCheckPerson();
                     if (bioPerson.canUse(person, false, true, userWTuserID)) {
                         const biography = new Biography(theSourceRules);
@@ -4351,7 +786,7 @@ export class CC7 {
         $("#getDegreeButton").prop("disabled", true);
         $("#cancelLoad").show();
         CC7.cancelLoadController = new AbortController();
-        CC7.showShakingTree();
+        Utils.showShakingTree();
         const wtId = wtViewRegistry.getCurrentWtId();
         const degreeCounts = {};
         const [resultByKey, isPartial, actualMaxDegree] = await CC7.makePagedCallAndAddPeople(
@@ -4361,7 +796,7 @@ export class CC7 {
         );
         if (resultByKey == "aborted") {
             wtViewRegistry.showWarning("Profile retrieval cancelled.");
-            CC7.hideShakingTree();
+            Utils.hideShakingTree();
         } else {
             if (isPartial) {
                 wtViewRegistry.showWarning(
@@ -4376,7 +811,7 @@ export class CC7 {
             const [nrDirectAncestors, nrDuplicateAncestors] = CC7.categoriseProfiles(root, maxWantedDegree);
 
             window.cc7Degree = Math.min(maxWantedDegree, actualMaxDegree);
-            CC7.hideShakingTree();
+            Utils.hideShakingTree();
             if ($("#degreesTable").length != 0) {
                 $("#degreesTable").remove();
                 $("#ancReport").remove();
@@ -4398,8 +833,8 @@ export class CC7 {
                         ).toFixed(2)}%) occur more than once due to pedigree collapse.</p>`
                 )
             );
-            CC7.buildDegreeTableData(degreeCounts, 1, window.cc7Degree);
-            CC7.addPeopleTable(CC7.tableCaption(window.rootId, window.cc7Degree, !CC7.ONE_DEGREE));
+            CC7.buildDegreeTableData(degreeCounts, 1);
+            PeopleTable.addPeopleTable(PeopleTable.tableCaption());
         }
         $("#getPeopleButton").prop("disabled", false);
         $("#getDegreeButton").prop("disabled", false);
@@ -4427,7 +862,11 @@ export class CC7 {
         if ($("#degreesTable").length == 0) {
             $("#cc7Container").append(
                 $(
-                    "<table id='degreesTable'><tr><th colspan=2>Collecting Profiles</th></tr><tr><th>Data Request No.</th></tr><tr><th>Received</th></tr><tr><th>Total</th></tr></table>"
+                    "<table id='degreesTable'>" +
+                        "<tr><th colspan=2>Collecting Profiles</th></tr>" +
+                        "<tr><th>Data Request No.</th></tr>" +
+                        "<tr><th>Received</th></tr>" +
+                        "<tr><th>Total</th></tr></table>"
                 )
             );
         }
@@ -4509,13 +948,13 @@ export class CC7 {
         try {
             const result = await WikiTreeAPI.postToAPI(
                 {
-                    appId: CC7.APP_ID,
+                    appId: Settings.APP_ID,
                     action: "getPeople",
                     keys: ids.join(","),
                     start: start,
                     limit: limit,
                     nuclear: depth,
-                    fields: CC7.currentSettings["biocheck_options_biocheckOn"]
+                    fields: Settings.current["biocheck_options_biocheckOn"]
                         ? CC7.GET_PEOPLE_FIELDS + ",Bio"
                         : CC7.GET_PEOPLE_FIELDS,
                 },
@@ -4537,7 +976,7 @@ export class CC7 {
         const ABOVE = true;
         const BELOW = false;
         const collator = new Intl.Collator();
-        CC7.fixBirthDate(theRoot);
+        Utils.fixBirthDate(theRoot);
         theRoot.isAncestor = false;
         theRoot.isAbove = false;
         // Note: unlike the usual case, where the queue contains nodes still to be "visited" and processed,
@@ -4577,7 +1016,7 @@ export class CC7 {
                 const person = window.people.get(pId);
                 if (person) {
                     if (typeof person.fixedBirthDate === "undefined") {
-                        CC7.fixBirthDate(person);
+                        Utils.fixBirthDate(person);
                     }
                     // Add this person's relatives to the queue
                     const rels = firstIteration
@@ -4644,7 +1083,7 @@ export class CC7 {
                 const person = window.people.get(pId);
                 if (person) {
                     if (typeof person.fixedBirthDate === "undefined") {
-                        CC7.fixBirthDate(person);
+                        Utils.fixBirthDate(person);
                     }
                     // Add this person's relatives to the queue
                     const rels = firstIteration
@@ -4669,7 +1108,7 @@ export class CC7 {
             }
             firstIteration = false;
         }
-        console.log(`nr direct ancestor profiles=${directAncestors.size}, nr dups=${duplicates.size}`, duplicates);
+        console.log(`nr direct ancestor profiles=${directAncestors.size}, nr dups=${duplicates.size}`);
         return [nrProfiles, duplicates.size];
 
         function setAndShouldAdd(pId, where) {
@@ -4698,31 +1137,6 @@ export class CC7 {
         }
     }
 
-    static fixBirthDate(person) {
-        let birthDate = CC7.ymdFix(person.BirthDate);
-        if (birthDate == "" && person.BirthDateDecade) {
-            birthDate = person.BirthDateDecade || "";
-        }
-        person.fixedBirthDate = birthDate;
-    }
-
-    static makeFilename() {
-        return (
-            CC7.makeSheetname() +
-            new Date().toISOString().replace("T", "_").replaceAll(":", "-").slice(0, 19) +
-            "_" +
-            $("#cc7Subset").val().substring(0, 3) +
-            (CC7.anyFilterActive() ? "_filtered" : "")
-        );
-    }
-
-    static makeSheetname() {
-        const theDegree = $("#cc7Degree").val();
-        let fileName = `CC${theDegree}_${wtViewRegistry.getCurrentWtId()}_`;
-        fileName += $("div.degreeView").length ? `Deg_${theDegree}_` : "";
-        return fileName;
-    }
-
     static downloadArray(array, fileName) {
         // Convert the JavaScript array to a string
         const arrayString = JSON.stringify(array);
@@ -4737,7 +1151,7 @@ export class CC7 {
 
         // Append the link to the DOM and trigger the download
         document.body.appendChild(link);
-        link.trigger("click");
+        link.click();
 
         // Remove the link from the DOM
         document.body.removeChild(link);
@@ -4749,9 +1163,9 @@ export class CC7 {
             .querySelector("#peopleTable")
             ?.querySelectorAll(".filter-input")
             .forEach((input, inputIndex) => {
-                input.removeEventListener("input", CC7.stripLtGt);
-                input.removeEventListener("input", CC7.filterListener);
-                input.removeEventListener("click", CC7.clearFilterClickListener);
+                input.removeEventListener("input", PeopleTable.stripLtGt);
+                input.removeEventListener("input", PeopleTable.filterListener);
+                input.removeEventListener("click", PeopleTable.clearFilterClickListener);
             });
         $("#cc7PBFilter").off("select2:select");
 
@@ -4776,8 +1190,19 @@ export class CC7 {
     static handleFileDownload() {
         try {
             CC7.collapsePeople();
-            const fileName = CC7.makeFilename();
-            CC7.downloadArray([[window.rootId, window.cc7Degree], ...window.people.entries()], fileName);
+            const fileName = PeopleTable.makeFilename();
+            CC7.downloadArray(
+                [
+                    [
+                        window.rootId,
+                        window.cc7Degree,
+                        $(wtViewRegistry.WT_ID_TEXT).val(),
+                        $("#cc7Container").hasClass("degreeView"),
+                    ],
+                    ...window.people.entries(),
+                ],
+                fileName
+            );
         } finally {
             CC7.expandPeople(window.people);
         }
@@ -4828,69 +1253,82 @@ export class CC7 {
             return;
         }
 
+        let isOneDegree = false;
         const reader = new FileReader();
         reader.onload = async function (e) {
             const contents = e.target.result;
             try {
                 let oldFormat = false;
                 const peeps = JSON.parse(contents);
-                const [r, x] = peeps.shift();
+                const [rId, cc7Deg, rWtId, oneDeg] = peeps.shift();
                 [window.people, oldFormat] = CC7.expandPeople(new Map(peeps));
-                window.rootId = r;
-                window.cc7Degree = Number.isFinite(x) ? (window.cc7Degree = x) : 0;
+                window.rootId = rId;
+                window.cc7Degree = Number.isFinite(cc7Deg) ? cc7Deg : 0;
+                isOneDegree = oneDeg;
                 if (oldFormat) {
                     wtViewRegistry.showNotice(
                         "This file is still in an old format. It is recommended that you regenerate it " +
                             "after reloading profiles from WikiTree."
                     );
                 }
+                const root = window.people.get(window.rootId);
+                $(wtViewRegistry.WT_ID_TEXT).val(root?.Name || rWtId || window.rootId);
             } catch (error) {
-                CC7.hideShakingTree();
+                Utils.hideShakingTree();
                 wtViewRegistry.showError(`The input file is not valid: ${error}`);
                 return;
             }
 
-            const root = window.people.get(window.rootId);
-            $(wtViewRegistry.WT_ID_TEXT).val(root.Name);
             const degreeCounts = {};
             let maxDegree = 0;
+            let minDegree = 1000;
             for (const aPerson of window.people.values()) {
                 if (aPerson.Hide) continue;
-                if (degreeCounts[aPerson.Meta.Degrees]) {
-                    degreeCounts[aPerson.Meta.Degrees] = degreeCounts[aPerson.Meta.Degrees] + 1;
+                const pDeg = aPerson.Meta.Degrees;
+                if (degreeCounts[pDeg]) {
+                    degreeCounts[pDeg] = degreeCounts[pDeg] + 1;
                 } else {
-                    degreeCounts[aPerson.Meta.Degrees] = 1;
+                    degreeCounts[pDeg] = 1;
                 }
-                if (aPerson.Meta.Degrees > maxDegree) {
-                    maxDegree = aPerson.Meta.Degrees;
+                if (pDeg > maxDegree) {
+                    maxDegree = pDeg;
+                }
+                if (pDeg < minDegree) {
+                    minDegree = pDeg;
                 }
             }
-            if (window.cc7Degree == 0) window.cc7Degree = maxDegree;
-            const theDegree = Math.min(window.cc7Degree, CC7.MAX_DEGREE);
-            CC7.hideShakingTree();
-            CC7.addPeopleTable(CC7.tableCaption(window.rootId, theDegree, !CC7.ONE_DEGREE));
+            isOneDegree ||= minDegree != 0;
+            if (window.cc7Degree == 0) window.cc7Degree = Math.min(maxDegree - 1, CC7.MAX_DEGREE);
+            // const theDegree = Math.min(window.cc7Degree, CC7.MAX_DEGREE);
+            if (isOneDegree) {
+                $("#cc7Container").addClass("degreeView");
+            } else {
+                $("#cc7Container").removeClass("degreeView");
+            }
+            Utils.hideShakingTree();
+            PeopleTable.addPeopleTable(PeopleTable.tableCaption());
             $("#cc7Container #cc7Subset").before(
                 $(
                     "<table id='degreesTable'>" +
                         "<tr id='trDeg'><th>Degrees</th></tr>" +
                         "<tr id='trCon'><th>Connections</th></tr>" +
-                        "<tr id='trTot'><th>Total</th></tr></table>"
+                        (isOneDegree ? "" : "<tr id='trTot'><th>Total</th></tr></table>")
                 )
             );
-            CC7.buildDegreeTableData(degreeCounts, 1, theDegree);
+            CC7.buildDegreeTableData(degreeCounts, isOneDegree ? window.cc7Degree : 1);
             CC7.handleDegreeChange(window.cc7Degree);
         };
 
         try {
             CC7.clearDisplay();
-            CC7.showShakingTree(() => reader.readAsText(file));
+            Utils.showShakingTree(() => reader.readAsText(file));
         } catch (error) {
-            CC7.hideShakingTree();
+            Utils.hideShakingTree();
             wtViewRegistry.showError(`The input file is not valid: ${error}`);
         }
     }
 
-    static buildDegreeTableData(degreeCounts, fromDegree, toDegree) {
+    static buildDegreeTableData(degreeCounts, fromDegree) {
         function addTableCol(i, degreeSum) {
             $("#trDeg").append($(`<td>${i}</td>`));
             $("#trCon").append($(`<td>${degreeCounts[i]}</td>`));
@@ -4899,7 +1337,7 @@ export class CC7 {
             }
         }
         let degreeSum = 0;
-        for (let i = fromDegree; i <= toDegree; ++i) {
+        for (let i = fromDegree; i <= window.cc7Degree; ++i) {
             degreeSum = degreeSum + degreeCounts[i];
             addTableCol(i, degreeSum);
         }
@@ -4907,189 +1345,5 @@ export class CC7 {
             degreeSum = degreeSum + degreeCounts[-1];
             addTableCol(-1, degreeSum);
         }
-    }
-
-    static cc7excelOut(fileType) {
-        const bioCheck = CC7.currentSettings["biocheck_options_biocheckOn"];
-        const idMap = new Map();
-        const sheetName = CC7.makeSheetname();
-
-        const wb = XLSX.utils.book_new();
-        wb.Props = {
-            Title: sheetName,
-            Subject: sheetName,
-            Author: "WikiTree",
-            CreatedDate: new Date(),
-        };
-
-        wb.SheetNames.push(sheetName);
-        const ws_data = [];
-        const headings = [
-            "WikiTree ID",
-            "Degree",
-            "Parents",
-            "Siblings",
-            "Spouses",
-            "Children",
-            "Given Name(s)",
-            "Last name at birth",
-            "Current last name",
-            "Birth date",
-            "Birth place",
-            "Death date",
-            "Death place",
-            "Age",
-            "Manager",
-            "Created",
-            "Modified",
-        ];
-        if (bioCheck) {
-            // Add biocheck column
-            headings.splice(1, 0, "Bio Issue");
-        }
-
-        ws_data.push(headings, []);
-        $("#peopleTable > tbody tr").each(function () {
-            const row = $(this);
-            if (!row.is(":visible")) return;
-            const tds = row.find("td");
-            let birthdate, birthplace, deathdate, deathplace, deathAge, created, touched, bioIssue;
-            tds.each(function () {
-                if ($(this).hasClass("privBio")) {
-                    bioIssue = $(this).hasClass("bioIssue");
-                }
-                if ($(this).hasClass("birthdate")) {
-                    birthdate = $(this).text();
-                }
-                if ($(this).hasClass("birthlocation")) {
-                    birthplace = $(this).text();
-                }
-                if ($(this).hasClass("deathdate")) {
-                    deathdate = $(this).text();
-                }
-                if ($(this).hasClass("deathlocation")) {
-                    deathplace = $(this).text();
-                }
-                if ($(this).hasClass("age-at-death")) {
-                    deathAge = $(this).text();
-                }
-                if ($(this).hasClass("created")) {
-                    created = $(this).text();
-                }
-                if ($(this).hasClass("touched")) {
-                    touched = $(this).text();
-                }
-            });
-            const pData = [
-                row.data("name"),
-                row.data("degree"),
-                row.data("parent"),
-                row.data("sibling"),
-                row.data("spouse"),
-                row.data("child"),
-                row.data("firstname"),
-                row.data("lnab"),
-                row.data("lnc"),
-                birthdate,
-                birthplace,
-                deathdate,
-                deathplace,
-                deathAge,
-                row.data("manager"),
-                created,
-                touched,
-            ];
-            if (bioCheck) {
-                // Add biocheck column and create id lookup table
-                pData.splice(1, 0, bioIssue);
-                if (bioIssue) idMap.set(row.data("name"), row.data("id"));
-            }
-            ws_data.push(pData);
-        });
-
-        const ws = XLSX.utils.aoa_to_sheet(ws_data);
-        Object.getOwnPropertyNames(ws)
-            .filter((k) => k.startsWith("A"))
-            .forEach((aCell) => {
-                const bCell = `B${aCell.substring(1)}`;
-                const wtId = ws[aCell].v;
-                if (wtId.match(/.+\-.+/)) {
-                    // Add a hyperlink to the WtId cell
-                    ws[aCell].l = { Target: `https://www.wikitree.com/wiki/${wtId}` };
-                }
-                if (bioCheck && ws[bCell].v === true) {
-                    // Add a BioCeck column with the BioCheck report as comment
-                    const id = idMap.get(wtId);
-                    const person = window.people.get(id);
-                    if (person?.bioCheckReport) {
-                        ws[bCell].c = [{ a: "WT", t: formCommentFromReport(person.bioCheckReport) }];
-                    }
-                }
-            });
-
-        function formCommentFromReport(bioCheckReport) {
-            let comment = "";
-            let cnt = 0;
-            for (let [msg, subLines] of bioCheckReport) {
-                if (subLines && subLines.length > 0) {
-                    let subList = "";
-                    for (const line of subLines) {
-                        subList = subList.concat("\n  * ", line);
-                    }
-                    msg += subList;
-                }
-                if (++cnt == 1) {
-                    comment += `${cnt}. `;
-                } else {
-                    comment += `\n${cnt}. `;
-                }
-                comment += msg;
-            }
-            return comment;
-        }
-
-        wb.Sheets[sheetName] = ws;
-
-        function s2ab(s) {
-            const buf = new ArrayBuffer(s.length);
-            const view = new Uint8Array(buf);
-            for (let i = 0; i < s.length; i++) view[i] = s.charCodeAt(i) & 0xff;
-            return buf;
-        }
-
-        const wscols = [
-            { wch: 20 },
-            { wch: 5 },
-            { wch: 5 },
-            { wch: 5 },
-            { wch: 5 },
-            { wch: 5 },
-            { wch: 25 },
-            { wch: 20 },
-            { wch: 20 },
-            { wch: 10 },
-            { wch: 40 },
-            { wch: 10 },
-            { wch: 40 },
-            { wch: 5 },
-            { wch: 10 },
-            { wch: 10 },
-        ];
-        if (bioCheck) {
-            wscols.splice(1, 0, { wch: 5 });
-        }
-
-        ws["!cols"] = wscols;
-
-        const wbout = XLSX.write(wb, { bookType: fileType, type: "binary" });
-        saveAs(new Blob([s2ab(wbout)], { type: "application/octet-stream" }), CC7.makeFilename() + "." + fileType);
-    }
-
-    static getCookie(name) {
-        return WikiTreeAPI.cookie(name) || null;
-    }
-
-    static setCookie(name, value, options) {
-        return WikiTreeAPI.cookie(name, value, options);
     }
 }
